### 2021-03-14 #3102 @whedon
[PRE REVIEW]: Semi-Automatic Classification Plugin: A Python tool for the download and processing of remote sensing images in QGIS
**Submitting author:** @semiautomaticgit (<a href="http://orcid.org/0000-0001-7283-116X">Luca Congedo</a>)
**Repository:** <a href="https://github.com/semiautomaticgit/SemiAutomaticClassificationPlugin" target ="_blank">https://github.com/semiautomaticgit/SemiAutomaticClassificationPlugin</a>
**Version:** 7.8.0
