### 2021-03-14 #3102 @arfon
:wave: @bstabler - would you be able to edit this paper for JOSS?
