+++ dock/scpdock.py
@@ -299,6 +299,7 @@ class SCPDock:
 		dT = cfg.utls.getTime()
 		cfg.inptDir = cfg.tmpDir + '/' + name + dT
 		oDir = cfg.utls.makeDirectory(cfg.inptDir)
+		nm = ''
 		# unzip to temp dir
 		try:
 			with cfg.zipfileSCP.ZipFile(shapeFilePath) as zOpen:
@@ -319,7 +320,9 @@ class SCPDock:
 			return 'No'
 		# try to remove SCP input
 		try:
-			cfg.utls.removeLayer(name)
+			# issue in QGIS 3.18
+			#cfg.utls.removeLayer(name)
+			pass
 		except:
 			pass
 		# convert to geopackage
@@ -330,6 +333,11 @@ class SCPDock:
 				nm = nm2
 			except:
 				pass
+		if not cfg.osSCP.path.isfile(cfg.inptDir + '/' + nm):
+			cfg.mx.msgErr59()
+			# logger
+			cfg.utls.logCondition(str(__name__) + '-' + (cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), 'Error training input')
+			return 'No'
 		try:
 			tSS = cfg.utls.addVectorLayer(cfg.inptDir + '/' + nm)
 		except:
@@ -435,7 +443,13 @@ class SCPDock:
 			return 'No'
 		tSS = cfg.shpLay
 		# create memory layer
-		provider = tSS.dataProvider()
+		try:
+			provider = tSS.dataProvider()
+		except Exception as err:
+			cfg.mx.msgErr59()
+			# logger
+			cfg.utls.logCondition(str(__name__) + '-' + (cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), 'Error training input')
+			return 'No'
 		fields = provider.fields().toList()
 		pCrs = cfg.utls.getCrs(tSS)
 		mL = cfg.qgisCoreSCP.QgsVectorLayer('MultiPolygon?crs=' + str(pCrs.toWkt()), cfg.trnLay, 'memory')
@@ -599,7 +613,7 @@ class SCPDock:
 			return 'No'
 		# try to remove SCP input
 		try:
-			cfg.utls.removeLayer(cfg.trnLay)
+			cfg.utls.removeLayerByLayer(cfg.shpLay)
 		except:
 			pass
 		# create memory layer
@@ -1666,17 +1680,15 @@ class SCPDock:
 	def showHideROI(self):
 		try:
 			if cfg.show_ROI_radioButton.isChecked():
-				l = cfg.utls.selectLayerbyName(cfg.trnLay)
-				if l is not None:
-					cfg.utls.setLayerVisible(l, True)
-				cfg.utls.moveLayerTop(l)
+				if cfg.shpLay is not None:
+					cfg.utls.setLayerVisible(cfg.shpLay, True)
+				cfg.utls.moveLayerTop(cfg.shpLay)
 				cfg.rbbrBndPol.show()
 				# ROI point
 				self.vx.show()
 			else:
-				l = cfg.utls.selectLayerbyName(cfg.trnLay)
-				if l is not None:
-					cfg.utls.setLayerVisible(l, False)
+				if cfg.shpLay is not None:
+					cfg.utls.setLayerVisible(cfg.shpLay, False)
 				cfg.rbbrBndPol.hide()
 				# ROI point
 				self.vx.hide()
@@ -2335,7 +2347,14 @@ class SCPDock:
 			except Exception as err:
 				# logger
[…]
