+++ i18n/semiautomaticclassificationplugin_es.ts
@@ -5479,7 +5479,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation>Desconocido</translation>
     </message>
     <message>
-        <location filename="../maininterface/clusteringTab.py" line="1254"/>
+        <location filename="../maininterface/zonalStatRasterTab.py" line="255"/>
         <source>Class</source>
         <translation>Clase</translation>
     </message>
