### 2021-03-31 #3102 @kthyng
@joferkington Are you interested in reviewing this submission for JOSS? The reviews are done all through github following checklists and opening issues in the software repo. We ask for reviews to be finished in 4-6 weeks but it tends to be an iterative process so it would be best to start the process earlier in that window. You can get more information here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
