+++ i18n/semiautomaticclassificationplugin_de.ts
@@ -6065,7 +6065,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="9232"/>
+        <location filename="../core/utils.py" line="9234"/>
         <source>SCP: completed process</source>
         <translation type="unfinished"></translation>
     </message>
