+++ i18n/semiautomaticclassificationplugin_fr.ts
@@ -5657,7 +5657,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="9232"/>
+        <location filename="../core/utils.py" line="9234"/>
         <source>SCP: completed process</source>
         <translation type="unfinished"></translation>
     </message>
