+++ maininterface/bandcombination.py
@@ -50,11 +50,15 @@ class BandCombination:
 	# cross classification calculation
 	def bandSetCombination(self, batch = 'No', bandSet = None, rasterOutput = None):
 		if batch == 'No':
-			combRstPath = cfg.utls.getSaveFileName(None, cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Save band combination raster output'), '', '*.tif', 'tif')
+			combRstPath = cfg.utls.getSaveFileName(None, cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Save band combination raster output'), '', 'TIF file (*.tif);;VRT file (*.vrt)')
 		else:
 			combRstPath = rasterOutput
+		# virtual raster
+		vrtR = 'No'
 		if combRstPath is not False:
-			if combRstPath.lower().endswith('.tif'):
+			if combRstPath.lower().endswith('.vrt'):
+				vrtR = 'Yes'
+			elif combRstPath.lower().endswith('.tif'):
 				pass
 			else:
 				combRstPath = combRstPath + '.tif'
@@ -92,25 +96,32 @@ class BandCombination:
 				cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), " Warning")
 				return 'No'
 			cfg.uiUtls.updateBar(10)
-			rEPSG = cfg.utls.getEPSGRaster(cfg.bndSetLst[0])				
+			NoDataVal = cfg.NoDataVal
+			rCrs = cfg.utls.getCrsGDAL(cfg.bndSetLst[0])
+			rEPSG = cfg.osrSCP.SpatialReference()
+			rEPSG.ImportFromWkt(rCrs)
 			if rEPSG is None:
 				if batch == 'No':
 					cfg.uiUtls.removeProgressBar()
 				cfg.mx.msgWar28()
 				# logger
-				cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), " Warning")
-				return 'No'	
+				cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), ' Warning')
+				return 'No'
 			cfg.uiUtls.updateBar(20)
-			NoDataVal = cfg.NoDataVal				
-			for b in range(0, len(cfg.bndSetLst)):						
-				EPSG = cfg.utls.getEPSGRaster(cfg.bndSetLst[b])
-				if str(EPSG) != str(rEPSG):
+			for b in range(0, len(cfg.bndSetLst)):
+				eCrs = cfg.utls.getCrsGDAL(cfg.bndSetLst[b])
+				EPSG = cfg.osrSCP.SpatialReference()
+				EPSG.ImportFromWkt(eCrs)
+				if EPSG.IsSame(rEPSG) != 1:
 					if cfg.bandSetsList[bandSetNumber][0] == 'Yes':
 						nD = cfg.utls.imageNoDataValue(cfg.bndSetLst[b])
 						if nD is None:
 							nD = NoDataVal
-						tPMD = cfg.utls.createTempRasterPath('tif')
-						cfg.utls.GDALReprojectRaster(cfg.bndSetLst[b], tPMD, "GTiff", None, "EPSG:" + str(rEPSG), "-ot Float32 -dstnodata " + str(nD))
+						tPMD = cfg.utls.createTempRasterPath('vrt')
[…]
