+++ i18n/semiautomaticclassificationplugin_es.ts
@@ -5449,12 +5449,12 @@ Do you want to add the required fields to this shapefile?</source>
         <translation>Guardar ráster</translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7130"/>
+        <location filename="../core/utils.py" line="7136"/>
         <source>Build overviews</source>
         <translation>Construir vistas generales</translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7130"/>
+        <location filename="../core/utils.py" line="7136"/>
         <source>Do you want to build the external overviews of bands?</source>
         <translation>¿Quieres construir vistas generales externas?</translation>
     </message>
@@ -5489,7 +5489,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation>Porcentaje %</translation>
     </message>
     <message>
-        <location filename="../maininterface/clipmultiplerasters.py" line="114"/>
+        <location filename="../maininterface/clipmultiplerasters.py" line="105"/>
         <source>Select a directory where to save clipped rasters</source>
         <translation>Selecciona un directorio para guardar los rásters cortados</translation>
     </message>
@@ -6069,7 +6069,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished">Sentinel-3</translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="9216"/>
+        <location filename="../core/utils.py" line="9222"/>
         <source>SCP: completed process</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6484,27 +6484,27 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="6024"/>
+        <location filename="../core/utils.py" line="6030"/>
         <source>Calculating signature</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7398"/>
+        <location filename="../core/utils.py" line="7404"/>
         <source>Writing file</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="5258"/>
+        <location filename="../core/utils.py" line="5264"/>
         <source>Conversion to vector</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7079"/>
+        <location filename="../core/utils.py" line="7085"/>
         <source>Sieve</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7209"/>
+        <location filename="../core/utils.py" line="7215"/>
         <source>Building overviews</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6544,12 +6544,12 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="8663"/>
+        <location filename="../core/utils.py" line="8669"/>
         <source>Remove rows</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="8663"/>
+        <location filename="../core/utils.py" line="8669"/>
         <source>Are you sure you want to remove highlighted rows from the table?</source>
         <translation type="unfinished"></translation>
     </message>
