+++ i18n/semiautomaticclassificationplugin_es.ts
@@ -623,7 +623,7 @@ p, li { white-space: pre-wrap; }
         <translation>Previa </translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="23688"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="23698"/>
         <source>Semi-Automatic Classification Plugin</source>
         <translation>Semi-Automatic Classification Plugin</translation>
     </message>
@@ -658,12 +658,12 @@ p, li { white-space: pre-wrap; }
         <translation>Solo con vista previa</translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="22395"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="22405"/>
         <source>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;&lt;span &gt;Run&lt;/span&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
         <translation>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;&lt;span &gt;Ejecutar&lt;/span&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="23521"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="23531"/>
         <source>Import library</source>
         <translation>Importar librería</translation>
     </message>
@@ -743,7 +743,7 @@ p, li { white-space: pre-wrap; }
         <translation>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Máximo porcentaje de cobertura de nubes&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="21005"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="21015"/>
         <source>Filter</source>
         <translation>Filtro</translation>
     </message>
@@ -753,12 +753,12 @@ p, li { white-space: pre-wrap; }
         <translation>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;Filtrar imágenes&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="21479"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="21489"/>
         <source>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p &gt;Delete row&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
         <translation>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p &gt;Suprimir fila&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="22242"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="22252"/>
         <source>Plot</source>
         <translation>Gráfico</translation>
     </message>
@@ -768,7 +768,7 @@ p, li { white-space: pre-wrap; }
[…]
