+++ core/config.py
@@ -794,7 +794,7 @@ functionNames.append([[' cloud_masking', 'cfg.batchT.performCloudMasking',  'cfg
 functionNames.append([[' goes_conversion', 'cfg.batchT.performGOESConversion', 'cfg.goesT.GOES', ['input_dir : \'\'', 'use_nodata : 1', 'nodata_value : 0', 'create_bandset : 1', 'output_dir : \'\'', 'band_set : 1']]])
 functionNames.append([[' landsat_conversion', 'cfg.batchT.performLandsatConversion',  'cfg.landsatT.landsat', ['input_dir : \'\'', 'mtl_file_path : \'\'', 'celsius_temperature : 0', 'apply_dos1 : 0', 'use_nodata : 1', 'nodata_value : 0', 'pansharpening : 0', 'create_bandset : 1', 'output_dir : \'\'', 'band_set : 1']]])
 functionNames.append([[' modis_conversion', 'cfg.batchT.performMODISConversion', 'cfg.MODIST.MODIS', ['input_raster_path : \'\'', 'reproject_wgs84 : 1', 'use_nodata : 1', 'nodata_value : -999', 'create_bandset : 1', 'output_dir : \'\'', 'band_set : 1']]])
-functionNames.append([[' mosaic_bandsets', 'cfg.batchT.performMosaicBandSets', 'cfg.mosaicBS.mosaicBandSets', ['band_set_list : \'\'', 'output_dir : \'\'', 'output_name_prefix : \'mosaic\'']]])
+functionNames.append([[' mosaic_bandsets', 'cfg.batchT.performMosaicBandSets', 'cfg.mosaicBS.mosaicBandSets', ['band_set_list : \'\'', 'virtual_output : 0', 'output_dir : \'\'', 'output_name_prefix : \'mosaic\'']]])
 functionNames.append([[' neighbor_pixels', 'cfg.batchT.performNeighborPixels', 'cfg.clssNghbr.classNeighbor', ['band_set : 1', 'matrix_size : 1', 'matrix_file_path : \'\'', 'output_name_prefix : \'neighbor\'', 'statistic : \'sum\'', 'stat_value : 50', 'output_dir : \'\'']]])
 functionNames.append([[' reproject_raster_bands', 'cfg.batchT.performReprojectRasters',  'cfg.rprjRstBndsT.reprojectRasters', ['band_set : 1', 'output_dir : \'\'', 'align_raster_path : \'\'', 'same_extent_reference : 0', 'epsg : \'\'','x_resolution : \'\'', 'y_resolution : \'\'', 'resample_pixel_factor : \'\'',  'resampling_method : \'near\'', 'output_nodata_value : -32768', 'data_type : \'auto\'',  'output_name_prefix : \'reproj\'']]])
 functionNames.append([[' sentinel1_conversion', 'cfg.batchT.performSentinel1Conversion', 'cfg.sentinel1T.sentinel1', ['input_raster_path : \'\'', 'xml_file_path : \'\'', 'vh : 1', 'vv : 1',  'raster_project : 0',  'raster_projections_band_set : 1', 'convert_to_db : 1', 'use_nodata : 1', 'nodata_value : 0', 'create_bandset : 1', 'output_dir : \'\'', 'band_set : 1']]])
