+++ core/config.py
@@ -782,8 +782,8 @@ functionNames.append([[' add_new_bandset', 'cfg.batchT.performAddNewBandSet',  '
 functionNames.append([[' add_raster', 'cfg.batchT.performAddRaster',  'cfg.utls.addRasterOrBand', ['input_raster_path : \'\'', 'input_raster_name : \'\'', 'band_set : 1', 'center_wavelength : 1']]])
 functionNames.append([[' create_bandset', 'cfg.batchT.performBandSetCreation',  'cfg.bst.addFileToBandSet', ['raster_path_list : \'\'', 'center_wavelength : \'\'', 'wavelength_unit : 1', 'multiplicative_factor : \'\'', 'additive_factor : \'\'', 'date : \'\'']]])
 functionNames.append([[' open_training_input', 'cfg.batchT.performOpenTrainingInput',  'cfg.SCPD.openInput', ['training_file_path : \'\'']]])
-functionNames.append([[' remove_band_from_bandset', 'cfg.batchT.performRemoveBandFromBandSet',  'cfg.bst.removeBandsFromBandSet', ['band_set : 1', 'band_list : \'\'']]])
-functionNames.append([[' remove_bandset', 'cfg.batchT.performRemoveBandSet',  'cfg.bst.removeBandSetTab', ['band_set : 1']]])
+functionNames.append([[' remove_band_from_bandset', 'cfg.batchT.performRemoveBandFromBandSet',  'cfg.bst.removeBandsFromBandSet', ['band_set : 1', 'band_list : \'\'', 'unload_bands : 0']]])
+functionNames.append([[' remove_bandset', 'cfg.batchT.performRemoveBandSet',  'cfg.bst.removeBandSetTab', ['band_set : 1', 'unload_bands : 0']]])
 functionNames.append([[' select_bandset', 'cfg.batchT.performBandSetSelection',  'cfg.bst.selectBandSetTab', ['band_set : 1']]])
 functionNames.append([['Band calculation']])
 functionNames.append([[' band_calc', 'cfg.batchT.performBandCalc',  'cfg.bCalc.calculate', ['expression : \'\'', 'output_raster_path : \'\'', 'extent_same_as_raster_name : \'\'', 'align : 1', 'extent_intersection : 1', 'input_nodata_as_value : 0', 'use_value_nodata : 0', 'output_nodata_value : -32768', 'data_type : \'Float32\'', 'scale_value : 1',  'offset_value : 0', 'band_set : 1']]])
@@ -794,7 +794,7 @@ functionNames.append([[' cloud_masking', 'cfg.batchT.performCloudMasking',  'cfg
 functionNames.append([[' goes_conversion', 'cfg.batchT.performGOESConversion', 'cfg.goesT.GOES', ['input_dir : \'\'', 'use_nodata : 1', 'nodata_value : 0', 'create_bandset : 1', 'output_dir : \'\'', 'band_set : 1']]])
 functionNames.append([[' landsat_conversion', 'cfg.batchT.performLandsatConversion',  'cfg.landsatT.landsat', ['input_dir : \'\'', 'mtl_file_path : \'\'', 'celsius_temperature : 0', 'apply_dos1 : 0', 'use_nodata : 1', 'nodata_value : 0', 'pansharpening : 0', 'create_bandset : 1', 'output_dir : \'\'', 'band_set : 1']]])
[…]
