+++ maininterface/batchTab.py
@@ -523,7 +523,7 @@ class BatchTab:
 			pName = pSplit[0].lower().replace(' ', '')
 			# input file path inside ' '
 			if pName == 'input_raster_path':
-				pSplitX = pSplit[1]
+				pSplitX = pSplit[1].strip()
 				if cfg.workingDir is not None:
 					pSplitX = pSplitX.replace(cfg.workingDirNm, cfg.workingDir)
 				g = cfg.reSCP.findall('[\'](.*?)[\']',pSplitX.replace('\\', '/'))
@@ -532,7 +532,7 @@ class BatchTab:
 				else:
 					return 'No', pName
 			elif pName == 'input_raster_name':
-				pSplitX = pSplit[1]
+				pSplitX = pSplit[1].strip()
 				g = cfg.reSCP.findall('[\'](.*?)[\']',pSplitX.replace('\\', '/'))
 				if len(g[0].strip()) > 0:
 					name = '\'' + g[0].strip() + '\''
@@ -540,12 +540,12 @@ class BatchTab:
 					return 'No', pName
 			elif pName == 'band_set':
 				try:
-					bandset = str(int(eval(pSplit[1].replace(' ', ''))) - 1)
+					bandset = str(int(eval(pSplit[1].strip().replace(' ', ''))) - 1)
 				except:
 					return 'No', pName
 			elif pName == 'center_wavelength':
 				try:
-					wavelength = str(float(pSplit[1].replace(' ', '')))
+					wavelength = str(float(pSplit[1].strip().replace(' ', '')))
 				except:
 					return 'No', pName
 			else:
@@ -570,7 +570,7 @@ class BatchTab:
 			pName = pSplit[0].lower().replace(' ', '')
 			# input directory inside ' '
 			if pName == 'input_dir':
-				pSplitX = pSplit[1]
+				pSplitX = pSplit[1].strip()
 				if cfg.workingDir is not None:
 					pSplitX = pSplitX.replace(cfg.workingDirNm, cfg.workingDir)
 				g = cfg.reSCP.findall('[\'](.*?)[\']',pSplitX.replace('\\', '/'))
@@ -584,7 +584,7 @@ class BatchTab:
 					return 'No', pName
 			# output directory inside ' '
 			elif pName == 'output_dir':
-				pSplitX = pSplit[1]
+				pSplitX = pSplit[1].strip()
 				if cfg.workingDir is not None:
 					pSplitX = pSplitX.replace(cfg.workingDirNm, cfg.workingDir)
 				g = cfg.reSCP.findall('[\'](.*?)[\']',pSplitX.replace('\\', '/'))
@@ -595,60 +595,60 @@ class BatchTab:
 			# band set number
 			elif pName == 'band_set':
 				try:
-					bandset = str(int(eval(pSplit[1].replace(' ', ''))) - 1)
+					bandset = str(int(eval(pSplit[1].strip().replace(' ', ''))) - 1)
 				except:
 					return 'No', pName
 			# MTL file path inside ' '
 			elif pName == 'mtl_file_path':
-				pSplitX = pSplit[1]
+				pSplitX = pSplit[1].strip()
 				if cfg.workingDir is not None:
 					pSplitX = pSplitX.replace(cfg.workingDirNm, cfg.workingDir)
 				g = cfg.reSCP.findall('[\'](.*?)[\']',pSplitX.replace('\\', '/'))
 				cfg.ui.label_27.setText(g[0])
 			# temperature in Celsius checkbox (1 checked or 0 unchecked)
 			elif pName == 'celsius_temperature':
-				if pSplit[1].replace(' ', '') == '1':
+				if pSplit[1].strip().replace(' ', '') == '1':
[…]
