+++ i18n/semiautomaticclassificationplugin_fr.ts
@@ -5667,7 +5667,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="9225"/>
+        <location filename="../core/utils.py" line="9156"/>
         <source>SCP: completed process</source>
         <translation type="unfinished"></translation>
     </message>
@@ -5872,12 +5872,12 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7139"/>
+        <location filename="../core/utils.py" line="7070"/>
         <source>Build overviews</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7139"/>
+        <location filename="../core/utils.py" line="7070"/>
         <source>Do you want to build the external overviews of bands?</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6492,27 +6492,27 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="6033"/>
+        <location filename="../core/utils.py" line="5964"/>
         <source>Calculating signature</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7407"/>
+        <location filename="../core/utils.py" line="7338"/>
         <source>Writing file</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="5267"/>
+        <location filename="../core/utils.py" line="5265"/>
         <source>Conversion to vector</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7088"/>
+        <location filename="../core/utils.py" line="7019"/>
         <source>Sieve</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7218"/>
+        <location filename="../core/utils.py" line="7149"/>
         <source>Building overviews</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6552,12 +6552,12 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="8672"/>
+        <location filename="../core/utils.py" line="8603"/>
         <source>Remove rows</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="8672"/>
+        <location filename="../core/utils.py" line="8603"/>
         <source>Are you sure you want to remove highlighted rows from the table?</source>
         <translation type="unfinished"></translation>
     </message>
