+++ core/utils.py
@@ -2954,7 +2954,8 @@ class Utils:
 	def writeRasterBlock(self, gdalRaster, bandNumber, dataArray, pixelStartColumn, pixelStartRow, nodataValue=None):
 		b = gdalRaster.GetRasterBand(bandNumber)
 		y, x = dataArray.shape
-		b.WriteRaster(pixelStartColumn, pixelStartRow, x, y, dataArray.tostring())
+		#b.WriteRaster(pixelStartColumn, pixelStartRow, x, y, dataArray.tostring())
+		b.WriteArray(dataArray, pixelStartColumn, pixelStartRow)
 		if nodataValue is not None:
 			b.SetNoDataValue(nodataValue)
 		b.FlushCache()
@@ -4762,12 +4763,10 @@ class Utils:
 								pCount = 0
 								while True:
 									pCount = pCount + 1
-									#writeOut = cfg.utls.writeRaster(wrtFile, [tLX + x*pSX, pSX, rGT[2], tLY + y*pSY, rGT[4], pSY], rP, bSX, bSY, format, oo, outputNoData, scl, offs, compress, compressFormat, dataType)
 									writeOut = cfg.utls.writeRasterNew(wrtFile, x-roX, y-roY, bSX, bSY, oo, outputNoData, scl, offs, dataType)
 									oR = cfg.gdalSCP.Open(wrtFile, cfg.gdalSCP.GA_ReadOnly)
 									bO = oR.GetRasterBand(1)
 									oo2 = cfg.utls.readArrayBlock(bO, x-roX, y-roY, bSX, bSY)
-									#oo2 = bO.ReadAsArray()
 									bO = None
 									oR = None
 									checkOO = cfg.np.allclose(oo, oo2, equal_nan=True)
@@ -4778,7 +4777,6 @@ class Utils:
 									if pCount > 3:
 										break
 							else:
-								#writeOut = cfg.utls.writeRaster(wrtFile, [tLX + x*pSX, pSX, rGT[2], tLY + y*pSY, rGT[4], pSY], rP, bSX, bSY, format, oo, outputNoData, scl, offs, compress, compressFormat, dataType)
 								# logger
 								cfg.utls.logToFile(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), 'writeRasterNew compress ' + str(compress))
 								writeOut = cfg.utls.writeRasterNew(wrtFile, x-roX, y-roY, bSX, bSY, oo, outputNoData, scl, offs, dataType)
@@ -5396,78 +5394,10 @@ class Utils:
 				pass
 			cfg.utls.writeArrayBlock(r, 1, arrayFile, section[0], section[1])
 			r = None
-			
-	# write raster
-	def writeRaster(self, rasterPath, geoTransform, projection, bSX, bSY, format, dataArray, nodataValue = None, scale = None, offset = None, compress = 'No', compressFormat = 'LZW', dataType = None):
-		tD = cfg.gdalSCP.GetDriverByName('GTiff')
-		if compress == 'Yes':
-			option = ['COMPRESS=' + compressFormat]
-		else:
-			option = []
-		if dataType is not None:
-			try:
-				format = eval('cfg.gdalSCP.GDT_' + dataType)
-			except:
-				format = cfg.gdalSCP.GDT_Float32
-			if dataType == 'Float64':
-				oType = cfg.np.float64
-			elif dataType == 'Float32':
-				oType = cfg.np.float32
-			elif dataType == 'Int32':
-				oType = cfg.np.int32
[…]

[…]
