+++ ui/ui_semiautomaticclassificationplugin.py
@@ -5180,7 +5180,10 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.class_neighbor_threshold_spinBox.setObjectName("class_neighbor_threshold_spinBox")
         self.gridLayout_300.addWidget(self.class_neighbor_threshold_spinBox, 1, 1, 1, 1)
         spacerItem86 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.gridLayout_300.addItem(spacerItem86, 1, 2, 1, 1)
+        self.gridLayout_300.addItem(spacerItem86, 1, 3, 1, 1)
+        self.circular_structure_checkBox = QtWidgets.QCheckBox(self.tab_neighbor_pixels)
+        self.circular_structure_checkBox.setObjectName("circular_structure_checkBox")
+        self.gridLayout_300.addWidget(self.circular_structure_checkBox, 1, 2, 1, 1)
         self.gridLayout_240.addLayout(self.gridLayout_300, 1, 0, 1, 3)
         self.horizontalLayout_72 = QtWidgets.QHBoxLayout()
         self.horizontalLayout_72.setObjectName("horizontalLayout_72")
@@ -10287,6 +10290,8 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.label_280.setText(_translate("SemiAutomaticClassificationPlugin", "Neighbor distance in pixels"))
         self.band_set_comb_spinBox_15.setToolTip(_translate("SemiAutomaticClassificationPlugin", "<html><head/><body><p>Band set number</p></body></html>"))
         self.class_neighbor_threshold_spinBox.setToolTip(_translate("SemiAutomaticClassificationPlugin", "<html><head/><body><p>Distance in pixels</p></body></html>"))
+        self.circular_structure_checkBox.setToolTip(_translate("SemiAutomaticClassificationPlugin", "<html><head/><body><p>If checked, neighbor pixels are calculated inside a circle of radius equal to the distance in pixels</p></body></html>"))
+        self.circular_structure_checkBox.setText(_translate("SemiAutomaticClassificationPlugin", "Circular"))
         self.tabWidget_preprocessing.setTabText(self.tabWidget_preprocessing.indexOf(self.tab_neighbor_pixels), _translate("SemiAutomaticClassificationPlugin", "Neighbor pixels"))
         self.SCP_tabs.setTabText(self.SCP_tabs.indexOf(self.tab_preprocessing), _translate("SemiAutomaticClassificationPlugin", "Preprocessing"))
         self.label_253.setText(_translate("SemiAutomaticClassificationPlugin", " Run"))
