+++ maininterface/reprojectrasterbands.py
@@ -117,7 +117,7 @@ class ReprojectRasterBands:
 		self.reprojectRasters('No', None, alignRaster, sameExtent, EPSG, xResolution, yResolution, resPixelSize, resample, type, noData, outName, bandSet)
 		
 	# reproject multiple rasters
-	def reprojectRasters(self, batch = 'No', outputDirectory = None, alignRasterPath = None, sameExtent = 'No', EPSGCode = None, xResolution = None, yResolution = None, resamplePixelFactor = None, resamplingMethod = None, outputType = None, noDataValue = None, outputName = None, bandSetNumber = None):
+	def reprojectRasters(self, batch = 'No', outputDirectory = None, alignRasterPath = None, sameExtent = 'No', EPSGCode = None, xResolution = None, yResolution = None, resamplePixelFactor = None, resamplingMethod = None, outputType = None, noDataValue = None, outputName = None, bandSetNumber = None, outFormat = 'GTiff'):
 		if bandSetNumber is None:
 			bandSet = cfg.ui.band_set_comb_spinBox_14.value()
 			bandSetNumber = bandSet - 1
@@ -182,7 +182,7 @@ class ReprojectRasterBands:
 			lC = cfg.utls.selectLayerbyName(l, 'Yes')
 			bLS = cfg.utls.layerSource(lC)
 			# calculate minimal extent
-			if alignRasterPath is not None and sameExtent == 'No':
+			if alignRasterPath is not None:
 				# raster extent and pixel size
 				try:
 					leftS, rightS, topS, bottomS, pXS, pYS, rPS, unitS = cfg.utls.imageGeoTransform(bLS)
@@ -195,23 +195,26 @@ class ReprojectRasterBands:
 					cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), ' ERROR exception: ' + str(err))
 					cfg.mx.msgErr61(lC)
 					return 'No'
-				# minimum extent
-				if leftSP < left:
-					leftR = left - int(2 + (left - leftSP) / pX) * pX
+				if sameExtent == 'No':
+					# minimum extent
+					if leftSP < left:
+						leftR = left - int(2 + (left - leftSP) / pX) * pX
+					else:
+						leftR = left + int((leftSP- left) / pX - 2) * pX
+					if rightSP > right:
+						rightR = right + int(2 + (rightSP - right) / pX) * pX
+					else:
+						rightR = right - int((right - rightSP) / pX - 2) * pX
+					if topSP > top:
+						topR = top + int(2 + (topSP - top) / pY) * pY
+					else:
+						topR = top - int((top - topSP) / pY - 2) * pY
+					if bottomSP > bottom:
+						bottomR = bottom + int((bottomSP - bottom) / pY - 2) * pY
+					else:
+						bottomR = bottom - int(2 + (bottom - bottomSP) / pY) * pY
 				else:
-					leftR = left + int((leftSP- left) / pX - 2) * pX
-				if rightSP > right:
-					rightR = right + int(2 + (rightSP - right) / pX) * pX
-				else:
-					rightR = right - int((right - rightSP) / pX - 2) * pX
-				if topSP > top:
-					topR = top + int(2 + (topSP - top) / pY) * pY
-				else:
-					topR = top - int((top - topSP) / pY - 2) * pY
-				if bottomSP > bottom:
-					bottomR = bottom + int((bottomSP - bottom) / pY - 2) * pY
-				else:
[…]
