+++ maininterface/neighborpixelsTab.py
@@ -53,7 +53,7 @@ class NeighborPixels:
 		cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), str(m))
 		
 	# classification neighbor
-	def classNeighbor(self, batch = 'No', bandSetNumber = None, outputDirectory = None, size = None, structure = None, statName = None, statPerc = None, outputName = None, circularStructure = None):
+	def classNeighbor(self, batch = 'No', bandSetNumber = None, outputDirectory = None, size = None, structure = None, statName = None, statPerc = None, outputName = None, circularStructure = None, virtual = None):
 			if bandSetNumber is None:
 				bandSet = cfg.ui.band_set_comb_spinBox_15.value()
 				bandSetNumber = bandSet - 1
@@ -66,6 +66,9 @@ class NeighborPixels:
 				o = outputDirectory
 			if len(o) > 0:
 				cfg.utls.makeDirectory(o)
+				if virtual is None:
+					if cfg.ui.neighbor_virtual_checkBox.isChecked() is True:
+						virtual = 'Yes'
 				if outputName is None:
 					outputName = cfg.ui.neighbor_output_name_lineEdit.text()
 					if len(outputName) > 0:
@@ -138,7 +141,7 @@ class NeighborPixels:
 					ee = statNp.replace('array', 'A')
 					try:
 						statPerc = int(statPerc)
-						ee = ee.replace(cfg.statPerc, str(statPerc) + ', axis=2')
+						ee = ee.replace(cfg.statPerc, str(statPerc))
 					except:
 						pass
 				else:
@@ -164,29 +167,23 @@ class NeighborPixels:
 						# logger
 						if cfg.logSetVal == 'Yes': cfg.utls.logToFile(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), ' ERROR exception: ' + str(err))
 						return 'No'
-				if 'nansum' in ee:
-					additionalLayer = 2
-				else:
-					additionalLayer = structure.shape[0]*structure.shape[1]*0.8
 				# logger
 				cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), 'structure ' + str(structure))
 				for y in range(0, len(bndSetSources)):
 					if cfg.actionCheck == 'Yes':
 						input = bndSetSources[y]
 						nd = cfg.utls.imageNoDataValue(input)
-						dType = cfg.utls.getRasterDataTypeName(input)
-						tPMD = cfg.utls.createTempRasterPath('vrt')
+						additionalLayer = 3
+						if virtual == 'Yes':
+							outputRaster = o + '/' + outputName + cfg.utls.fileNameNoExt(bndSetSources[y]) + '.vrt'
+							#outputRaster = tPMD
+						else:
+							#tPMD = cfg.utls.createTempRasterPath('vrt')
+							outputRaster = o + '/' + outputName + cfg.utls.fileNameNoExt(bndSetSources[y]) + '.tif'
 						# process calculation
[…]
