+++ i18n/semiautomaticclassificationplugin_ar.ts
@@ -622,7 +622,7 @@ p, li { white-space: pre-wrap; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="23708"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="23718"/>
         <source>Semi-Automatic Classification Plugin</source>
         <translation type="unfinished"></translation>
     </message>
@@ -657,12 +657,12 @@ p, li { white-space: pre-wrap; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="22415"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="22425"/>
         <source>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;&lt;span &gt;Run&lt;/span&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="23541"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="23551"/>
         <source>Import library</source>
         <translation type="unfinished"></translation>
     </message>
@@ -742,7 +742,7 @@ p, li { white-space: pre-wrap; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="21025"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="21035"/>
         <source>Filter</source>
         <translation type="unfinished"></translation>
     </message>
@@ -752,12 +752,12 @@ p, li { white-space: pre-wrap; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="21499"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="21509"/>
         <source>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p &gt;Delete row&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="22262"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="22272"/>
         <source>Plot</source>
         <translation type="unfinished"></translation>
     </message>
@@ -767,7 +767,7 @@ p, li { white-space: pre-wrap; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="23407"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="23417"/>
         <source>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;&lt;span &gt;Reset&lt;/span&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
         <translation type="unfinished"></translation>
     </message>
@@ -817,17 +817,17 @@ p, li { white-space: pre-wrap; }
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="22787"/>
+        <location filename="../ui/ui_semiautomaticclassificationplugin.ui" line="22797"/>
[…]
