+++ core/config.py
@@ -352,15 +352,15 @@ bandCalcFunctionNames.append(['@', '@'])
 bandCalcFunctionNames.append(['Custom'])
 # zonal stat raster
 statisticList = []
-statisticList.append(['Sum', 'np.nansum(array)'])
-statisticList.append(['Max', 'np.nanmax(array)'])
-statisticList.append(['Min', 'np.nanmin(array)'])
 statisticList.append(['Count', 'np.count_nonzero(~np.isnan(array))'])
+statisticList.append(['Max', 'np.nanmax(array)'])
 statisticList.append(['Mean', 'np.nanmean(array)'])
 statisticList.append(['Median', 'np.nanmedian(array)'])
+statisticList.append(['Min', 'np.nanmin(array)'])
 statPerc = '@statPerc@'
 statisticList.append(['Percentile', 'np.nanpercentile(array, ' + statPerc + ')'])
 statisticList.append(['StandardDeviation', 'np.nanstd(array)'])
+statisticList.append(['Sum', 'np.nansum(array)'])
 # SCP kml name
 kmlNm = 'SCP_kml'
 # subprocesses dictionary
@@ -795,7 +795,7 @@ functionNames.append([[' goes_conversion', 'cfg.batchT.performGOESConversion', '
 functionNames.append([[' landsat_conversion', 'cfg.batchT.performLandsatConversion',  'cfg.landsatT.landsat', ['input_dir : \'\'', 'mtl_file_path : \'\'', 'celsius_temperature : 0', 'apply_dos1 : 0', 'use_nodata : 1', 'nodata_value : 0', 'pansharpening : 0', 'create_bandset : 1', 'output_dir : \'\'', 'band_set : 1']]])
 functionNames.append([[' modis_conversion', 'cfg.batchT.performMODISConversion', 'cfg.MODIST.MODIS', ['input_raster_path : \'\'', 'reproject_wgs84 : 1', 'use_nodata : 1', 'nodata_value : -999', 'create_bandset : 1', 'output_dir : \'\'', 'band_set : 1']]])
 functionNames.append([[' mosaic_bandsets', 'cfg.batchT.performMosaicBandSets', 'cfg.mosaicBS.mosaicBandSets', ['band_set_list : \'\'', 'use_nodata : 1', 'nodata_value : 0', 'virtual_output : 0', 'output_dir : \'\'', 'output_name_prefix : \'mosaic\'']]])
-functionNames.append([[' neighbor_pixels', 'cfg.batchT.performNeighborPixels', 'cfg.clssNghbr.classNeighbor', ['band_set : 1', 'matrix_size : 1', 'circular : 0', 'matrix_file_path : \'\'', 'output_name_prefix : \'neighbor\'', 'statistic : \'sum\'', 'stat_value : 50', 'output_dir : \'\'']]])
+functionNames.append([[' neighbor_pixels', 'cfg.batchT.performNeighborPixels', 'cfg.clssNghbr.classNeighbor', ['band_set : 1', 'matrix_size : 1', 'circular : 0', 'matrix_file_path : \'\'', 'virtual_output : 0', 'output_name_prefix : \'neighbor\'', 'statistic : \'sum\'', 'stat_value : 50', 'output_dir : \'\'']]])
[…]

[…]
