+++ maininterface/dilationTab.py
@@ -65,7 +65,7 @@ class DilationRaster:
 		self.dilationClassification()
 		
 	# dilation classification
-	def dilationClassification(self, batch = 'No', rasterInput = None, rasterOutput = None):
+	def dilationClassification(self, batch = 'No', rasterInput = None, rasterOutput = None, circularStructure = None):
 		# class value list
 		valueList = self.checkValueList()
 		if len(valueList) > 0:
@@ -111,16 +111,23 @@ class DilationRaster:
 					nd = cfg.utls.imageNoDataValue(input)
 					dType = cfg.utls.getRasterDataTypeName(input)
 					size =  cfg.ui.dilation_threshold_spinBox.value()
-					connect = cfg.ui.dilation_connection_combo.currentText()
-					struct = cfg.utls.create3x3Window(connect)
-					for s in range(0, size):
-						if vrtR == 'Yes':
-							tPMD = outputRaster
+					if circularStructure is None:
+						if cfg.ui.circular_structure_checkBox_2.isChecked():
+							circularStructure = 'Yes'
 						else:
-							tPMD = cfg.utls.createTempRasterPath('vrt')
-						# process calculation
-						o = cfg.utls.multiProcessRaster(rasterPath = input, functionBand = 'No', functionRaster = cfg.utls.rasterDilation, outputRasterList = [tPMD], functionBandArgument = struct, functionVariable = valueList, progressMessage = cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Dilation '), virtualRaster = 'Yes', compress = 'No', outputNoDataValue = nd, dataType = dType, boundarySize = 3)
-						input = tPMD
+							circularStructure = 'No'
+					if circularStructure == 'No':
+						structure = cfg.np.ones((size*2+1,size*2+1))
+					else:
+						structure = cfg.utls.createCircularStructure(size)
+					additionalLayer = 3
+					if vrtR == 'Yes':
+						tPMD = outputRaster
+					else:
+						tPMD = cfg.utls.createTempRasterPath('vrt')
+					# process calculation
+					o = cfg.utls.multiProcessRaster(rasterPath = input, functionBand = 'No', functionRaster = cfg.utls.rasterDilation, outputRasterList = [tPMD], functionBandArgument = structure, functionVariable = valueList, progressMessage = cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Dilation '), virtualRaster = 'Yes', compress = cfg.rasterCompression, outputNoDataValue = nd, dataType = dType, boundarySize = structure.shape[0]+1, additionalLayer = additionalLayer)
+					input = tPMD
 					if vrtR != 'Yes':
 						# copy raster
 						try:
