+++ i18n/semiautomaticclassificationplugin_fr.ts
@@ -5907,7 +5907,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/batchTab.py" line="3674"/>
+        <location filename="../maininterface/batchTab.py" line="3676"/>
         <source>missing parameter</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6912,7 +6912,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/neighborpixelsTab.py" line="184"/>
+        <location filename="../maininterface/neighborpixelsTab.py" line="182"/>
         <source>Neighbor pixels</source>
         <translation type="unfinished"></translation>
     </message>
