+++ i18n/semiautomaticclassificationplugin_pl.ts
@@ -5107,7 +5107,7 @@ ROI</source>
         <translation>Zapisz błąd macierzy wyjścia rastra</translation>
     </message>
     <message>
-        <location filename="../maininterface/crossclassificationTab.py" line="369"/>
+        <location filename="../maininterface/crossclassificationTab.py" line="381"/>
         <source>Classification</source>
         <translation>Klasyfikacja</translation>
     </message>
@@ -5117,7 +5117,7 @@ ROI</source>
         <translation>BłądKoduMacierzy</translation>
     </message>
     <message>
-        <location filename="../maininterface/crossclassificationTab.py" line="352"/>
+        <location filename="../maininterface/crossclassificationTab.py" line="364"/>
         <source>Reference</source>
         <translation>Odniesienie</translation>
     </message>
@@ -5197,12 +5197,12 @@ ROI</source>
         <translation>Zapisz raster</translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7151"/>
+        <location filename="../core/utils.py" line="7149"/>
         <source>Build overviews</source>
         <translation>Zbuduj podglądy</translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7151"/>
+        <location filename="../core/utils.py" line="7149"/>
         <source>Do you want to build the external overviews of bands?</source>
         <translation>Czy na pewno zbudować zewnętrzne podglądy na kanały?</translation>
     </message>
@@ -5964,7 +5964,7 @@ Chcesz dodać wymagane pola do tego shapefile?</translation>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/crossclassificationTab.py" line="367"/>
+        <location filename="../maininterface/crossclassificationTab.py" line="379"/>
         <source>CROSS MATRIX [</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6089,7 +6089,7 @@ Chcesz dodać wymagane pola do tego shapefile?</translation>
         <translation type="unfinished">Sentinel-3</translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="9239"/>
+        <location filename="../core/utils.py" line="9237"/>
         <source>SCP: completed process</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6174,12 +6174,12 @@ Chcesz dodać wymagane pola do tego shapefile?</translation>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/bandcombination.py" line="264"/>
+        <location filename="../maininterface/bandcombination.py" line="277"/>
         <source>RasterValue</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/bandcombination.py" line="264"/>
+        <location filename="../maininterface/bandcombination.py" line="277"/>
         <source>Combination</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6504,27 +6504,27 @@ Chcesz dodać wymagane pola do tego shapefile?</translation>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="6045"/>
+        <location filename="../core/utils.py" line="6043"/>
         <source>Calculating signature</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7419"/>
+        <location filename="../core/utils.py" line="7417"/>
         <source>Writing file</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="5346"/>
+        <location filename="../core/utils.py" line="5344"/>
         <source>Conversion to vector</source>
         <translation type="unfinished"></translation>
[…]
