+++ ui/ui_semiautomaticclassificationplugin.py
@@ -1311,7 +1311,7 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.gridLayout_175.addLayout(self.gridLayout_174, 0, 0, 1, 1)
         self.toolBox_4.addItem(self.page_6, "")
         self.page_9 = QtWidgets.QWidget()
-        self.page_9.setGeometry(QtCore.QRect(0, 0, 417, 169))
+        self.page_9.setGeometry(QtCore.QRect(0, 0, 417, 165))
         self.page_9.setObjectName("page_9")
         self.gridLayout_181 = QtWidgets.QGridLayout(self.page_9)
         self.gridLayout_181.setObjectName("gridLayout_181")
@@ -5316,7 +5316,7 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.toolBox_PCA.setStyleSheet("")
         self.toolBox_PCA.setObjectName("toolBox_PCA")
         self.page_16 = QtWidgets.QWidget()
-        self.page_16.setGeometry(QtCore.QRect(0, 0, 923, 413))
+        self.page_16.setGeometry(QtCore.QRect(0, 0, 459, 196))
         self.page_16.setObjectName("page_16")
         self.gridLayout_182 = QtWidgets.QGridLayout(self.page_16)
         self.gridLayout_182.setObjectName("gridLayout_182")
@@ -5429,7 +5429,7 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.toolBox_kmeans.setStyleSheet("")
         self.toolBox_kmeans.setObjectName("toolBox_kmeans")
         self.page_18 = QtWidgets.QWidget()
-        self.page_18.setGeometry(QtCore.QRect(0, 0, 923, 413))
+        self.page_18.setGeometry(QtCore.QRect(0, 0, 764, 390))
         self.page_18.setObjectName("page_18")
         self.gridLayout_152 = QtWidgets.QGridLayout(self.page_18)
         self.gridLayout_152.setObjectName("gridLayout_152")
@@ -6086,7 +6086,7 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.toolBox_random_forest.setStyleSheet("")
         self.toolBox_random_forest.setObjectName("toolBox_random_forest")
         self.page_21 = QtWidgets.QWidget()
-        self.page_21.setGeometry(QtCore.QRect(0, 0, 923, 413))
+        self.page_21.setGeometry(QtCore.QRect(0, 0, 634, 327))
         self.page_21.setObjectName("page_21")
         self.gridLayout_287 = QtWidgets.QGridLayout(self.page_21)
         self.gridLayout_287.setObjectName("gridLayout_287")
@@ -7780,30 +7780,11 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.erosion_threshold_spinBox.setProperty("value", 1)
         self.erosion_threshold_spinBox.setObjectName("erosion_threshold_spinBox")
         self.horizontalLayout_14.addWidget(self.erosion_threshold_spinBox)
+        self.circular_structure_checkBox_3 = QtWidgets.QCheckBox(self.tab_erosion)
+        self.circular_structure_checkBox_3.setObjectName("circular_structure_checkBox_3")
+        self.horizontalLayout_14.addWidget(self.circular_structure_checkBox_3)
         spacerItem139 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
         self.horizontalLayout_14.addItem(spacerItem139)
-        self.label_150 = QtWidgets.QLabel(self.tab_erosion)
[…]
