+++ core/utils.py
@@ -3193,7 +3193,7 @@ class Utils:
 			return 'No'
 		
 	# clip raster with another raster
-	def clipRasterByRaster(self, rasterClippedList, rasterClipping, outputRasterDir = None, outFormat = 'GTiff', nodataVal = None, progressMessage = 'Clip', stats = None, parallelWritingCheck = None, outputNameRoot = None):
+	def clipRasterByRaster(self, rasterClippedList, rasterClipping, outputRasterDir = None, outFormat = 'GTiff', nodataVal = None, progressMessage = 'Clip', stats = None, parallelWritingCheck = None, outputNameRoot = None, compress = 'No', compressFormat = 'LZW'):
 		tPMD = cfg.utls.createTempRasterPath('vrt')
 		bList = rasterClippedList.copy()
 		bList.append(rasterClipping)
@@ -3227,7 +3227,7 @@ class Utils:
 				functionList.append(e)
 				variableList.append(varList)
 			oM = cfg.utls.createTempRasterList(bC-1)
-			oMR = cfg.utls.createRasterFromReference(rD, 1, oM, cfg.NoDataVal, 'GTiff', cfg.rasterDataType, 0, None, 'No')
+			oMR = cfg.utls.createRasterFromReference(rD, 1, oM, cfg.NoDataVal, 'GTiff', cfg.rasterDataType, compress = compress, compressFormat = compressFormat)			
 			# close GDAL rasters
 			for b in range(0, len(oMR)):
 				oMR[b] = None
@@ -3769,23 +3769,45 @@ class Utils:
 		
 	# reclassify raster
 	def reclassifyRaster(self, gdalBandList, rasterSCPArrayfunctionBand, columnNumber, rowNumber, pixelStartColumn, pixelStartRow, outputArrayFile, functionBandArgument, functionVariableList, outputBandNumber):
-		# raster array
-		o = rasterSCPArrayfunctionBand
-		a = cfg.np.copy(o)
-		# logger
-		cfg.utls.logToFile(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), 'reclassifyRaster ' )
-		for i in functionBandArgument:
-			# create condition
-			try:
-				o[a==int(i[0])] = float(i[1])
-			except Exception as err:
+		try:
+			old, new = cfg.np.array(functionBandArgument).T
+			# logger
+			cfg.utls.logToFile(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), 'old ' + str(old))
+			cfg.utls.logToFile(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), 'new ' + str(new))
+			raster = cfg.np.nan_to_num(rasterSCPArrayfunctionBand[:,:,0])
+			if cfg.np.all(raster.astype(int) == raster) and cfg.np.all(old.astype(int) == old):
+				reclass = cfg.np.zeros(max(old.astype(int).max(), rasterSCPArrayfunctionBand.astype(int).max())+1) * cfg.np.nan
+				reclass[old.astype(int)] = new
+				o = reclass[raster.astype(int)].astype(cfg.np.float32)
+				o[cfg.np.isnan(o)] = raster[cfg.np.isnan(o)]
+				o[cfg.np.isnan(rasterSCPArrayfunctionBand[:,:,0])] = cfg.np.nan
+			else:
[…]
