+++ core/utils.py
@@ -2481,17 +2481,17 @@ class Utils:
 						pass
 					# set metadata xml
 					xml = '''
-					<SimpleSource>
+					<ComplexSource>
 					  <SourceFilename relativeToVRT='%i'>%s</SourceFilename>
 					  <SourceBand>%i</SourceBand>
 					  <SourceProperties RasterXSize='%i' RasterYSize='%i' DataType=%s BlockXSize='%i' BlockYSize='%i' />
 					  <SrcRect xOff='%i' yOff='%i' xSize='%i' ySize='%i' />
 					  <DstRect xOff='%i' yOff='%i' xSize='%i' ySize='%i' />
 					  <NODATA>%i</NODATA>
-					</SimpleSource>
+					</ComplexSource>
 					'''
 					source = xml % (relativeToVRT, source_path, bandNumber, gdalRaster2.RasterXSize, gdalRaster2.RasterYSize, 'Float32', x_block, y_block, xoffX, xoffY, gdalRaster2.RasterXSize, gdalRaster2.RasterYSize, offX, offY, rX2, rY2, noData)
-					band.SetMetadataItem('SimpleSource', source, 'new_vrt_sources')
+					band.SetMetadataItem('ComplexSource', source, 'new_vrt_sources')
 					if NoDataVal == 'Yes':
 						band.SetNoDataValue(noData)	
 					elif NoDataVal != 'No':
@@ -2580,20 +2580,20 @@ class Utils:
 							pass
 						# set metadata xml
 						xml = '''
-						<SimpleSource>
+						<ComplexSource>
 						  <SourceFilename relativeToVRT='%i'>%s</SourceFilename>
 						  <SourceBand>%i</SourceBand>
 						  <SourceProperties RasterXSize='%i' RasterYSize='%i' DataType=%s BlockXSize='%i' BlockYSize='%i' />
 						  <SrcRect xOff='%i' yOff='%i' xSize='%i' ySize='%i' />
 						  <DstRect xOff='%i' yOff='%i' xSize='%i' ySize='%i' />
 						  <NODATA>%i</NODATA>
-						</SimpleSource>
+						</ComplexSource>
 						'''
 						if aster == 'No':
 							source = xml % (relativeToVRT, source_path, bandNumber, gdalRaster2.RasterXSize, gdalRaster2.RasterYSize, 'Float32', x_block, y_block, xoffX, xoffY, gdalRaster2.RasterXSize, gdalRaster2.RasterYSize, offX, offY, rX2, rY2, noData)
 						else:
 							source = xml % (relativeToVRT, source_path, bandNumber, gdalRaster2.RasterXSize, gdalRaster2.RasterYSize, 'Float32', x_block, y_block, xoffX, xoffY, gdalRaster2.RasterXSize, gdalRaster2.RasterYSize, xoffX, xoffY, gdalRaster2.RasterXSize, gdalRaster2.RasterYSize, noData)
-						band.SetMetadataItem('SimpleSource', source, 'new_vrt_sources')
+						band.SetMetadataItem('ComplexSource', source, 'new_vrt_sources')
 						if NoDataVal == 'Yes':
 							band.SetNoDataValue(noData)	
 						elif NoDataVal != 'No':
@@ -2910,6 +2910,26 @@ class Utils:
 				bottomR = bottom
 			extra = '-tr ' + str(pX) + ' ' + str(pY) + ' -te ' + str(leftR) + ' ' + str(bottomR) + ' ' + str(rightR) + ' ' + str(topR)
[…]

[…]
