+++ maininterface/classSignatureTab.py
@@ -147,7 +147,7 @@ class ClassSignatureTab:
 			for x in sorted(cfg.parallelArrayDict):
 				try:
 					for ar in cfg.parallelArrayDict[x]:
-						values = cfg.np.append(values, ar[0, ::])
+						values = cfg.np.append(values, ar[0][0, ::])
 				except:
 					if batch == 'No':
 						cfg.utls.finishSound()
@@ -170,29 +170,29 @@ class ClassSignatureTab:
 			for x in sorted(cfg.parallelArrayDict):
 				try:
 					for ar in cfg.parallelArrayDict[x]:
-						for arX in ar:
+						for arX in ar[0]:
 							for i in range(0, len(bList) - 1):
 								for c in classes:
 									if 'COUNT_BAND_' + str(i) + '_c_' + str(c) == arX:
 										try:
-											cfg.rasterClassSignature['COUNT_BAND_' + str(i) + '_c_' + str(c)] = cfg.rasterClassSignature['COUNT_BAND_' + str(i) + '_c_' + str(c)] + ar[arX]
+											cfg.rasterClassSignature['COUNT_BAND_' + str(i) + '_c_' + str(c)] = cfg.rasterClassSignature['COUNT_BAND_' + str(i) + '_c_' + str(c)] + ar[0][arX]
 										except:
-											cfg.rasterClassSignature['COUNT_BAND_' + str(i) + '_c_' + str(c)] = ar[arX]
+											cfg.rasterClassSignature['COUNT_BAND_' + str(i) + '_c_' + str(c)] = ar[0][arX]
 									elif 'SUM_BAND_' + str(i) + '_c_' + str(c) == arX:
 										try:
-											cfg.rasterClassSignature['SUM_BAND_' + str(i) + '_c_' + str(c)] = cfg.rasterClassSignature['SUM_BAND_' + str(i) + '_c_' + str(c)] + ar[arX]
+											cfg.rasterClassSignature['SUM_BAND_' + str(i) + '_c_' + str(c)] = cfg.rasterClassSignature['SUM_BAND_' + str(i) + '_c_' + str(c)] + ar[0][arX]
 										except:
-											cfg.rasterClassSignature['SUM_BAND_' + str(i) + '_c_' + str(c)] = ar[arX]
+											cfg.rasterClassSignature['SUM_BAND_' + str(i) + '_c_' + str(c)] = ar[0][arX]
 									elif 'MINIMUM_BAND_' + str(i) + '_c_' + str(c) == arX:
 										try:
-											cfg.rasterClassSignature['MINIMUM_BAND_' + str(i) + '_c_' + str(c)] = min(ar[arX], cfg.rasterClassSignature['MINIMUM_BAND_' + str(i) + '_c_' + str(c)])
+											cfg.rasterClassSignature['MINIMUM_BAND_' + str(i) + '_c_' + str(c)] = min(ar[0][arX], cfg.rasterClassSignature['MINIMUM_BAND_' + str(i) + '_c_' + str(c)])
 										except:
-											cfg.rasterClassSignature['MINIMUM_BAND_' + str(i) + '_c_' + str(c)] = ar[arX]
+											cfg.rasterClassSignature['MINIMUM_BAND_' + str(i) + '_c_' + str(c)] = ar[0][arX]
 									elif 'MAXIMUM_BAND_' + str(i) + '_c_' + str(c) == arX:
 										try:
-											cfg.rasterClassSignature['MAXIMUM_BAND_' + str(i) + '_c_' + str(c)] = max(ar[arX], cfg.rasterClassSignature['MAXIMUM_BAND_' + str(i) + '_c_' + str(c)])
[…]
