+++ maininterface/crossclassificationTab.py
@@ -222,7 +222,7 @@ class CrossClassification:
 					for x in sorted(cfg.parallelArrayDict):
 						try:
 							for ar in cfg.parallelArrayDict[x]:
-								values = cfg.np.append(values, ar[0, ::])
+								values = cfg.np.append(values, ar[0][0, ::])
 						except:
 							if batch == 'No':
 								cfg.utls.finishSound()

[…]
