### 2021-06-06 #3172 @joferkington
Apologies for the mixup on my part, @kthyng, but it seems I waited to long to accept the invitation link: https://github.com/openjournals/joss-reviews/invitations  When I try I get an error with "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account."

As a result, I can't check off any items in the review checklist.

Any chance I could get an updated invitation to review?  My apologies again for the mixup!
