### 2021-06-20 #3172 @bstabler
Hi @semiautomaticgit.  Thank you for developing this powerful and extensive QGIS Semi-Automatic (Remote Sensing) Classification Plugin (SCP).  I was able to install the plugin and work my through [Tutorial 1](https://semiautomaticclassificationmanual-v5.readthedocs.io/en/latest/tutorial_1.html) and [Random Forest Classification](https://fromgistors.blogspot.com/2021/02/random-forest-classification-using-scp.html).  There is a lot of material here and so it takes a bit of time to get familiar with all the SCP plugin dialogs, settings, tools, etc.  You can see above that I checked all boxes except for two:

  - [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

According to your [development guidelines](https://github.com/semiautomaticgit/SemiAutomaticClassificationPlugin/blob/master/DEVELOPMENT.md#end-user-test
), it appears that there is no automatic/organized/suggested testing framework?  I think some kind of testing framework is needed for when people less familiar with the source code want to contribute and need a process for ensuring revisions not only work for their use case, but also work for other use cases (i.e. the revisions do not have unintended consequences).  Essentially there needs to be adequate test coverage.  I wonder if you can use the QGIS [unittesting framework](https://docs.qgis.org/3.16/en/docs/developers_guide/unittesting.html)?

  - [ ]  A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?

I think the statement of need really starts at line 61.  Before line 61, the section summarizes the need for remote sensing, which is good for the common reader, but I think it is not really needed (or could be condensed) since it is not the specific contribution of your work.  Your work is a convenience, easy-to-use, open source toolkit for doing remote sensing in QGIS that makes use of third party algorithms.

And one small comment for the paper, the Python [multiprocessing](https://docs.python.org/3/library/multiprocessing.html) library uses subprocesses instead of threads so I suggest revising line 18.

Thanks!
