### 2021-06-20 #3172 @semiautomaticgit
Hi @bstabler , thank you very much for reviewing my work!
I'm really glad that almost all the boxes were checked.
About your comments:

> According to your [development guidelines](https://github.com/semiautomaticgit/SemiAutomaticClassificationPlugin/blob/master/DEVELOPMENT.md#end-user-test), it appears that there is no automatic/organized/suggested testing framework? I think some kind of testing framework is needed for when people less familiar with the source code want to contribute and need a process for ensuring revisions not only work for their use case, but also work for other use cases (i.e. the revisions do not have unintended consequences). Essentially there needs to be adequate test coverage. I wonder if you can use the QGIS [unittesting framework](https://docs.qgis.org/3.16/en/docs/developers_guide/unittesting.html)?

Unfortunately I don't think I can use QGIS unit testing framework with my plugin.
At the moment there is a test unit for checking the correct installation of dependencies and basic functionalities (such as numpy, GDAL subprocess, internet connection), which can be run from the menu [Settings > Debug](https://semiautomaticclassificationmanual.readthedocs.io/en/latest/debugTab.html#settings-debug-tab).
Of course I'm planning to improve these test units also for other plugin functions.
Also, the steps for reporting a bug (activating the log file) are described in the documentation [documentation](https://semiautomaticclassificationmanual.readthedocs.io/en/latest/faq.html#how-can-i-report-an-error) .
Do you think it could be enough if I cite this in the paper?

> I think the statement of need really starts at line 61. Before line 61, the section summarizes the need for remote sensing, which is good for the common reader, but I think it is not really needed (or could be condensed) since it is not the specific contribution of your work. Your work is a convenience, easy-to-use, open source toolkit for doing remote sensing in QGIS that makes use of third party algorithms.

I'm going to reduce the length of introduction about remote sensing, and highlight the characteristics of my work.

> And one small comment for the paper, the Python [multiprocessing](https://docs.python.org/3/library/multiprocessing.html) library uses subprocesses instead of threads so I suggest revising line 18.

I'll correct this, and I'm going to submit a new version.
Thank you again for your time.
