### 2021-06-25 #3172 @bstabler
@semiautomaticgit - thanks for updating the paper.  @joferkington, @kthyng - what do you think about the automated tests item?
