+++ i18n/semiautomaticclassificationplugin_pl.ts
@@ -5213,12 +5213,12 @@ ROI</source>
         <translation>Zapisz raster</translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7276"/>
+        <location filename="../core/utils.py" line="7288"/>
         <source>Build overviews</source>
         <translation>Zbuduj podglądy</translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7276"/>
+        <location filename="../core/utils.py" line="7288"/>
         <source>Do you want to build the external overviews of bands?</source>
         <translation>Czy na pewno zbudować zewnętrzne podglądy na kanały?</translation>
     </message>
@@ -5293,57 +5293,57 @@ ROI</source>
         <translation>Zapisz listę punktów w pliku</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="251"/>
+        <location filename="../maininterface/settings.py" line="256"/>
         <source>Transparency </source>
         <translation>Przezroczystość</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="64"/>
+        <location filename="../maininterface/settings.py" line="69"/>
         <source>Save Log file</source>
         <translation>Zapisz Plik zdarzeń</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="191"/>
+        <location filename="../maininterface/settings.py" line="196"/>
         <source>Reset field names</source>
         <translation>Wyczyść nazwy pól</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="191"/>
+        <location filename="../maininterface/settings.py" line="196"/>
         <source>Are you sure you want to reset field names?</source>
         <translation>Czy na pewno wyczyścić nazwy pól?</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="202"/>
+        <location filename="../maininterface/settings.py" line="207"/>
         <source>Reset variable name</source>
         <translation>Wyczyść nazwę zmiennej</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="202"/>
+        <location filename="../maininterface/settings.py" line="207"/>
         <source>Are you sure you want to reset variable name?</source>
         <translation>Czy na pewno wyczyścić nazwę zmiennej?</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="210"/>
+        <location filename="../maininterface/settings.py" line="215"/>
         <source>Reset group name</source>
         <translation>Wyczyść nazwę grupy</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="210"/>
+        <location filename="../maininterface/settings.py" line="215"/>
         <source>Are you sure you want to reset group name?</source>
         <translation>Czy na pewno wyczyścić nazwę grupy?</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="218"/>
+        <location filename="../maininterface/settings.py" line="223"/>
         <source>Change temporary directory</source>
         <translation>Zmień tymczasowy katalog</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="218"/>
+        <location filename="../maininterface/settings.py" line="223"/>
         <source>Are you sure you want to change the temporary directory?</source>
         <translation>Czy na pewno zmienić tymczasowy katalog?</translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="236"/>
+        <location filename="../maininterface/settings.py" line="241"/>
         <source>Are you sure you want to reset the temporary directory?</source>
         <translation>Czy na pewno wyczyścić katalog tymczasowy?</translation>
[…]
