+++ i18n/semiautomaticclassificationplugin_fr.ts
@@ -5813,7 +5813,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/bandcombination.py" line="308"/>
+        <location filename="../maininterface/bandcombination.py" line="319"/>
         <source>RasterValue</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6943,7 +6943,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/bandcombination.py" line="265"/>
+        <location filename="../maininterface/bandcombination.py" line="276"/>
         <source>Cross classification </source>
         <translation type="unfinished"></translation>
     </message>
