### 2021-07-12 #3172 @kthyng
Sorry for my delay @bstabler and @semiautomaticgit. Automated tests are not necessary, but a user (or future developer) needs to be able [to test](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) the basic functionality of the software. This could be in the form of a notebook that steps through instructions to end up with a specified result that demonstrates the software is working correctly. Is at least that level of testing present?
