+++ i18n/semiautomaticclassificationplugin_pl.ts
@@ -4823,17 +4823,17 @@ ROI</source>
         <translation>Pomoc Online</translation>
     </message>
     <message>
-        <location filename="../maininterface/classificationTab.py" line="709"/>
+        <location filename="../maininterface/classificationTab.py" line="716"/>
         <source>Select a mask shapefile</source>
         <translation>Wybierz shapefile maski</translation>
     </message>
     <message>
-        <location filename="../maininterface/classificationTab.py" line="133"/>
+        <location filename="../maininterface/classificationTab.py" line="135"/>
         <source>Save classification output</source>
         <translation>Zapisz wyjście klasyfikacji</translation>
     </message>
     <message>
-        <location filename="../maininterface/classificationTab.py" line="755"/>
+        <location filename="../maininterface/classificationTab.py" line="762"/>
         <source>Select a qml style</source>
         <translation>Wybierz styl qml</translation>
     </message>
@@ -6040,7 +6040,7 @@ Chcesz dodać wymagane pola do tego shapefile?</translation>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/clusteringTab.py" line="806"/>
+        <location filename="../maininterface/clusteringTab.py" line="816"/>
         <source> Calculating. Please wait ...</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6080,22 +6080,22 @@ Chcesz dodać wymagane pola do tego shapefile?</translation>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/clusteringTab.py" line="1255"/>
+        <location filename="../maininterface/clusteringTab.py" line="1275"/>
         <source>Signature</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/clusteringTab.py" line="1263"/>
+        <location filename="../maininterface/clusteringTab.py" line="1283"/>
         <source>C_ID_</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/clusteringTab.py" line="829"/>
+        <location filename="../maininterface/clusteringTab.py" line="839"/>
         <source> Calculating classification. Please wait ...</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/clusteringTab.py" line="1253"/>
+        <location filename="../maininterface/clusteringTab.py" line="1273"/>
         <source>Clustering</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6215,7 +6215,7 @@ Chcesz dodać wymagane pola do tego shapefile?</translation>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/clusteringTab.py" line="1255"/>
+        <location filename="../maininterface/clusteringTab.py" line="1275"/>
         <source>Distance</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6475,7 +6475,7 @@ Chcesz dodać wymagane pola do tego shapefile?</translation>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/classificationTab.py" line="222"/>
+        <location filename="../maininterface/classificationTab.py" line="224"/>
         <source>Conversion to vector. Please wait ...</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6665,17 +6665,17 @@ Chcesz dodać wymagane pola do tego shapefile?</translation>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/clusteringTab.py" line="1103"/>
+        <location filename="../maininterface/clusteringTab.py" line="1123"/>
[…]
