+++ maininterface/clusteringTab.py
@@ -398,6 +398,7 @@ class ClusteringTab:
 					signatureList = []
 					kN = 1
 					classes = []
+					classesMP = []
 					for p in range(0, k):
 						sig = []
 						signature = []
@@ -424,6 +425,7 @@ class ClusteringTab:
 						s.append(0)
 						signatureList.append(s)
 						classes.append([kN, c])
+						classesMP.append([kN, None])
 						kN = kN + 1
 				else:
 					# random seeds
@@ -432,6 +434,7 @@ class ClusteringTab:
 					signatureList = []
 					kN = 1
 					classes = []
+					classesMP = []
 					for p in range(0, len(points)):
 						sig = cfg.utls.calculatePixelSignature(cfg.qgisCoreSCP.QgsPointXY(points[p][0], points[p][1]), cfg.bandSetsList[bandSetNumber][8], bandSetNumber, 'Pixel', 'No')
 						signatures1.append(sig)
@@ -454,6 +457,7 @@ class ClusteringTab:
 						s.append(0)
 						signatureList.append(s)
 						classes.append([kN, c])
+						classesMP.append([kN, None])
 						kN = kN + 1
 			except:
 				# seed signatures
@@ -466,6 +470,7 @@ class ClusteringTab:
 				signatures1 = k_or_sigs
 				signatureList = []
 				classes = []
+				classesMP = []
 				for p in range(0, k):
 					sig = signatures1[p][0]
 					signature = []
@@ -486,9 +491,14 @@ class ClusteringTab:
 					s.append(0)
 					signatureList.append(s)
 					classes.append([signatures1[p][1], signatures1[p][2]])
+					classesMP.append([signatures1[p][1], None])
 			# process calculation
 			classificationOptions = ['No', 'No', 'No', cfg.algBandWeigths, cfg.algThrshld]
-			o = cfg.utls.multiProcessRaster(rasterPath = rD, signatureList = signatureList, functionBand = 'Yes', functionRaster = cfg.utls.classificationMultiprocess, algorithmName = algorithmName, nodataValue = -999, macroclassCheck = 'No',classificationOptions = classificationOptions, functionBandArgument = cfg.multiAddFactorsVar, functionVariable = cfg.bandSetsList[bandSetNumber][6], progressMessage = cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Classification iteration ') + str(iteration + 1).replace('-1', '*').replace('0', '*'), virtualRaster = 'Yes', compress =  'No')	
+			# for multiprocess
+			signatureListMP = signatureList.copy()
+			for smp in signatureListMP:
+				smp[6] = None
+			o = cfg.utls.multiProcessRaster(rasterPath = rD, signatureList = signatureListMP, functionBand = 'Yes', functionRaster = cfg.utls.classificationMultiprocess, algorithmName = algorithmName, nodataValue = -999, macroclassCheck = 'No',classificationOptions = classificationOptions, functionBandArgument = cfg.multiAddFactorsVar, functionVariable = cfg.bandSetsList[bandSetNumber][6], progressMessage = cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Classification iteration ') + str(iteration + 1).replace('-1', '*').replace('0', '*'), virtualRaster = 'Yes', compress =  'No')	
 			if o == 'No':
 				return 'No', None, None
 			# output rasters
@@ -527,7 +537,7 @@ class ClusteringTab:
 			try:
 				# values finder
 				cfg.parallelArrayDict = {}
[…]
