### 2021-07-15 #3172 @semiautomaticgit
> Sorry for my delay @bstabler and @semiautomaticgit. Automated tests are not necessary, but a user (or future developer) needs to be able [to test](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) the basic functionality of the software. This could be in the form of a notebook that steps through instructions to end up with a specified result that demonstrates the software is working correctly. Is at least that level of testing present?

Thank you @kthyng , yes there is a testing procedure described in the [user manual ](https://semiautomaticclassificationmanual.readthedocs.io/en/latest/faq.html#how-can-i-report-an-error) which checks software dependencies and that basic functions work.
Also, there is a [basic tutorial](https://semiautomaticclassificationmanual.readthedocs.io/en/latest/tutorial_1.html) that one can follow to perform a classification and test that the software works.
Should I mention this in the paper?

Thank you again
