+++ i18n/semiautomaticclassificationplugin_fr.ts
@@ -5693,7 +5693,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="9394"/>
+        <location filename="../core/utils.py" line="9397"/>
         <source>SCP: completed process</source>
         <translation type="unfinished"></translation>
     </message>
@@ -5893,12 +5893,12 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7294"/>
+        <location filename="../core/utils.py" line="7297"/>
         <source>Build overviews</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7294"/>
+        <location filename="../core/utils.py" line="7297"/>
         <source>Do you want to build the external overviews of bands?</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6513,27 +6513,27 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="6188"/>
+        <location filename="../core/utils.py" line="6191"/>
         <source>Calculating signature</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7568"/>
+        <location filename="../core/utils.py" line="7571"/>
         <source>Writing file</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="5482"/>
+        <location filename="../core/utils.py" line="5485"/>
         <source>Conversion to vector</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7243"/>
+        <location filename="../core/utils.py" line="7246"/>
         <source>Sieve</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7373"/>
+        <location filename="../core/utils.py" line="7376"/>
         <source>Building overviews</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6573,12 +6573,12 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="8841"/>
+        <location filename="../core/utils.py" line="8844"/>
         <source>Remove rows</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="8841"/>
+        <location filename="../core/utils.py" line="8844"/>
         <source>Are you sure you want to remove highlighted rows from the table?</source>
         <translation type="unfinished"></translation>
     </message>
