### 2021-07-19 #3172 @kthyng
@semiautomaticgit The first link seems to point to error reporting, but not basic testing of the software — am I misinterpreting? Basically the user needs to be able to run the software through some basic steps and get a verified result to know it is working correctly.
