### 2021-07-19 #3172 @semiautomaticgit
Thank you @kthyng

> @semiautomaticgit The first link seems to point to error reporting, but not basic testing of the software — am I misinterpreting? Basically the user needs to be able to run the software through some basic steps and get a verified result to know it is working correctly.

Yes it is error reporting, but I was particularly referring to the point 3 of the instructions where it is described the button "Test dependencies" to check software dependencies, which should report if there are any installations issues.

The second link (tutorial) is a full basic tutorial that one can follow to perform a semi-automatic land cover classification. Also other tutorials are available in the user manual about other topics such as [Unsupervised Classification](https://semiautomaticclassificationmanual.readthedocs.io/en/latest/tutorial_2.html) and  [Random ForestClassification](https://semiautomaticclassificationmanual.readthedocs.io/en/latest/thematic_tutorial_random_forest.html).

Thank you again.
