+++ maininterface/landsatTab.py
@@ -38,7 +38,7 @@ cfg = __import__(str(__name__).split('.')[0] + '.core.config', fromlist=[''])
 class LandsatTab:
 
 	def __init__(self):
-		pass
+		self.nodataCalc = cfg.NoDataValUInt16 * 0.00001
 		
 	# landsat input
 	def inputLandsat(self):
@@ -282,7 +282,7 @@ class LandsatTab:
 					elif str(nm[len(nm) - 8: len(nm) - 1]) != '6_VCID_' and nm[len(nm) - 1].isdigit():
 						e = self.landsat8reflectance(sat, str(nm[len(nm) - 1]), REFLECTANCE_MULT, REFLECTANCE_ADD, RADIANCE_MULT, RADIANCE_ADD, RADIANCE_MAXIMUM, REFLECTANCE_MAXIMUM)
 						# band list
-						if int(nm[len(nm) - 1]) in [2, 3, 4, 5, 6, 7]:
+						if int(nm[len(nm) - 1]) in [1, 2, 3, 4, 5, 6, 7]:
 							bandSetList.append(int(nm[len(nm) - 1]) - 1)
 							bandSetNameList.append(oNm)
 							if int(nm[len(nm) - 1]) < 5:
@@ -340,9 +340,9 @@ class LandsatTab:
 			# open input with GDAL
 			rD = cfg.gdalSCP.Open(tPMDN, cfg.gdalSCP.GA_ReadOnly)
 			# output rasters
-			cfg.utls.createRasterFromReference(rD, 1, oM, cfg.NoDataVal, 'GTiff', cfg.rasterDataType, 0,  None, compress = cfg.rasterCompression, compressFormat = 'LZW')
+			cfg.utls.createRasterFromReference(rD, 1, oM, cfg.NoDataValUInt16, 'GTiff', 'UInt16', 0,  None, compress = cfg.rasterCompression, compressFormat = 'DEFLATE')
 			rD = None
-			o = cfg.utls.multiProcessRaster(rasterPath = tPMDN, functionBand = 'No', functionRaster = cfg.utls.calculateRaster, outputRasterList = oM, nodataValue = NoData, functionBandArgument = argumentList, functionVariable = variableList, progressMessage = cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Conversion'), parallel = cfg.parallelRaster)
+			o = cfg.utls.multiProcessRaster(rasterPath = tPMDN, functionBand = 'No', functionRaster = cfg.utls.calculateRaster, outputRasterList = oM, nodataValue = NoData, functionBandArgument = argumentList, functionVariable = variableList, progressMessage = cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Conversion'), parallel = cfg.parallelRaster, outputNoDataValue = cfg.NoDataValUInt16, scale = 0.00001, offset = 0)
 			if cfg.actionCheck == 'Yes':
 				for t in range(0, len(outputRasterList)):
 					cfg.shutilSCP.move(oM[t], outputRasterList[t])
@@ -398,6 +398,7 @@ class LandsatTab:
 				argumentList = panList[1].replace('LDNm', str(LDNmList[0]))
 			else:
 				argumentList = panList[1]
+			# band conversion
[…]
