+++ i18n/semiautomaticclassificationplugin_fr.ts
@@ -5693,7 +5693,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="9397"/>
+        <location filename="../core/utils.py" line="9420"/>
         <source>SCP: completed process</source>
         <translation type="unfinished"></translation>
     </message>
@@ -5893,12 +5893,12 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7297"/>
+        <location filename="../core/utils.py" line="7320"/>
         <source>Build overviews</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7297"/>
+        <location filename="../core/utils.py" line="7320"/>
         <source>Do you want to build the external overviews of bands?</source>
         <translation type="unfinished"></translation>
     </message>
@@ -5948,7 +5948,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/classreportTab.py" line="127"/>
+        <location filename="../maininterface/classreportTab.py" line="130"/>
         <source>Percentage %</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6513,27 +6513,27 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="6191"/>
+        <location filename="../core/utils.py" line="6214"/>
         <source>Calculating signature</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7571"/>
+        <location filename="../core/utils.py" line="7594"/>
         <source>Writing file</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="5485"/>
+        <location filename="../core/utils.py" line="5508"/>
         <source>Conversion to vector</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7246"/>
+        <location filename="../core/utils.py" line="7269"/>
         <source>Sieve</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7376"/>
+        <location filename="../core/utils.py" line="7399"/>
         <source>Building overviews</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6573,12 +6573,12 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="8844"/>
+        <location filename="../core/utils.py" line="8867"/>
         <source>Remove rows</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="8844"/>
+        <location filename="../core/utils.py" line="8867"/>
         <source>Are you sure you want to remove highlighted rows from the table?</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6618,7 +6618,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/landsatTab.py" line="436"/>
+        <location filename="../maininterface/landsatTab.py" line="437"/>
         <source>Pansharpening</source>
         <translation type="unfinished"></translation>
     </message>
