+++ core/config.py
@@ -680,7 +680,7 @@ resampling_methods = ['nearest_neighbour', 'average', 'sum', 'maximum', 'minimum
 NoSatellite = 'Band order'
 satGeoEye1 = 'GeoEye-1 [bands 1, 2, 3, 4]'
 satGOES = 'GOES [bands 1, 2, 3, 4, 5, 6]'
-satLandsat8 = 'Landsat 8 OLI [bands 2, 3, 4, 5, 6, 7]'
+satLandsat8 = 'Landsat 8 OLI [bands 1, 2, 3, 4, 5, 6, 7]'
 satLandsat7 = 'Landsat 7 ETM+ [bands 1, 2, 3, 4, 5, 7]'
 satLandsat45 = 'Landsat 4-5 TM [bands 1, 2, 3, 4, 5, 7]'
 satLandsat13 = 'Landsat 1-3 MSS [bands 4, 5, 6, 7]'
