+++ maininterface/landsatTab.py
@@ -38,7 +38,7 @@ cfg = __import__(str(__name__).split('.')[0] + '.core.config', fromlist=[''])
 class LandsatTab:
 
 	def __init__(self):
-		self.nodataCalc = cfg.NoDataValUInt16 * 0.00001
+		self.nodataCalc = cfg.NoDataValUInt16 * 0.0001
 		
 	# landsat input
 	def inputLandsat(self):
@@ -342,7 +342,7 @@ class LandsatTab:
 			# output rasters
 			cfg.utls.createRasterFromReference(rD, 1, oM, cfg.NoDataValUInt16, 'GTiff', 'UInt16', 0,  None, compress = cfg.rasterCompression, compressFormat = 'DEFLATE')
 			rD = None
-			o = cfg.utls.multiProcessRaster(rasterPath = tPMDN, functionBand = 'No', functionRaster = cfg.utls.calculateRaster, outputRasterList = oM, nodataValue = NoData, functionBandArgument = argumentList, functionVariable = variableList, progressMessage = cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Conversion'), parallel = cfg.parallelRaster, outputNoDataValue = cfg.NoDataValUInt16, scale = 0.00001, offset = 0)
+			o = cfg.utls.multiProcessRaster(rasterPath = tPMDN, functionBand = 'No', functionRaster = cfg.utls.calculateRaster, outputRasterList = oM, nodataValue = NoData, functionBandArgument = argumentList, functionVariable = variableList, progressMessage = cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Conversion'), parallel = cfg.parallelRaster, outputNoDataValue = cfg.NoDataValUInt16, scale = 0.0001, offset = 0)
 			if cfg.actionCheck == 'Yes':
 				for t in range(0, len(outputRasterList)):
 					cfg.shutilSCP.move(oM[t], outputRasterList[t])
