+++ ui/ui_semiautomaticclassificationplugin.py
@@ -1311,7 +1311,7 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.gridLayout_175.addLayout(self.gridLayout_174, 0, 0, 1, 1)
         self.toolBox_4.addItem(self.page_6, "")
         self.page_9 = QtWidgets.QWidget()
-        self.page_9.setGeometry(QtCore.QRect(0, 0, 417, 169))
+        self.page_9.setGeometry(QtCore.QRect(0, 0, 417, 165))
         self.page_9.setObjectName("page_9")
         self.gridLayout_181 = QtWidgets.QGridLayout(self.page_9)
         self.gridLayout_181.setObjectName("gridLayout_181")
@@ -5288,7 +5288,7 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.gridLayout_330.addLayout(self.horizontalLayout_34, 0, 1, 1, 1)
         self.toolBox_band_set_combination.addItem(self.page_29, "")
         self.page_30 = QtWidgets.QWidget()
-        self.page_30.setGeometry(QtCore.QRect(0, 0, 90, 90))
+        self.page_30.setGeometry(QtCore.QRect(0, 0, 98, 90))
         self.page_30.setObjectName("page_30")
         self.gridLayout_334 = QtWidgets.QGridLayout(self.page_30)
         self.gridLayout_334.setObjectName("gridLayout_334")
@@ -5316,7 +5316,7 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.toolBox_PCA.setStyleSheet("")
         self.toolBox_PCA.setObjectName("toolBox_PCA")
         self.page_16 = QtWidgets.QWidget()
-        self.page_16.setGeometry(QtCore.QRect(0, 0, 535, 213))
+        self.page_16.setGeometry(QtCore.QRect(0, 0, 459, 196))
         self.page_16.setObjectName("page_16")
         self.gridLayout_182 = QtWidgets.QGridLayout(self.page_16)
         self.gridLayout_182.setObjectName("gridLayout_182")
@@ -5401,7 +5401,7 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.gridLayout_182.addLayout(self.gridLayout_173, 1, 0, 1, 1)
         self.toolBox_PCA.addItem(self.page_16, "")
         self.page_17 = QtWidgets.QWidget()
-        self.page_17.setGeometry(QtCore.QRect(0, 0, 90, 90))
+        self.page_17.setGeometry(QtCore.QRect(0, 0, 98, 90))
         self.page_17.setObjectName("page_17")
         self.gridLayout_200 = QtWidgets.QGridLayout(self.page_17)
         self.gridLayout_200.setObjectName("gridLayout_200")
@@ -5429,7 +5429,7 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.toolBox_kmeans.setStyleSheet("")
         self.toolBox_kmeans.setObjectName("toolBox_kmeans")
         self.page_18 = QtWidgets.QWidget()
-        self.page_18.setGeometry(QtCore.QRect(0, 0, 764, 407))
+        self.page_18.setGeometry(QtCore.QRect(0, 0, 764, 390))
         self.page_18.setObjectName("page_18")
         self.gridLayout_152 = QtWidgets.QGridLayout(self.page_18)
         self.gridLayout_152.setObjectName("gridLayout_152")
@@ -5674,7 +5674,7 @@ class Ui_SemiAutomaticClassificationPlugin(object):
         self.gridLayout_152.addLayout(self.gridLayout_231, 3, 0, 1, 1)
[…]
