### 2021-08-19 #3172 @kthyng
@semiautomaticgit and @bstabler - Testing dependencies is important but not enough for a user to be sure the software is functioning correctly. Would a user, going through the tutorials, be able to recreate exactly the same conditions as shown in the tutorials such that they could be sure the code is performing properly? As in, they can run with the exact same inputs, and if they return the exact same outputs, they can know it runs correctly?
