### 2021-08-20 #3172 @semiautomaticgit
> @semiautomaticgit and @bstabler - Testing dependencies is important but not enough for a user to be sure the software is functioning correctly. Would a user, going through the tutorials, be able to recreate exactly the same conditions as shown in the tutorials such that they could be sure the code is performing properly? As in, they can run with the exact same inputs, and if they return the exact same outputs, they can know it runs correctly?

Thank you very much @kthyng .
Yes, a user is able to recreate the same conditions shown in the tutorial and check that the code is performing properly.
I have updated the beginning of the [tutorial](https://semiautomaticclassificationmanual.readthedocs.io/en/latest/tutorial_1.html) to clearly state that this can be used to verify that the software in functioning correctly, providing the same input data (i.e. the image and training sample) and output result (i.e. the classification).
I'm looking forward to receiving your comments.
