### 2021-08-20 #3172 @semiautomaticgit
> * [ ]   A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
>
> I think the statement of need really starts at line 61. Before line 61, the section summarizes the need for remote sensing, which is good for the common reader, but I think it is not really needed (or could be condensed) since it is not the specific contribution of your work. Your work is a convenience, easy-to-use, open source toolkit for doing remote sensing in QGIS that makes use of third party algorithms.
>
> And one small comment for the paper, the Python [multiprocessing](https://docs.python.org/3/library/multiprocessing.html) library uses subprocesses instead of threads so I suggest revising line 18.
>
> Thanks!

I have also updated the paper according to the indications from @bstabler .
Thank you again!
