### 2021-08-21 #3172 @bstabler
@kthyng - I agree with @semiautomaticgit that a user can (probably/typically) verify the performance of the software with the tutorial.  The tutorial is pretty comprehensive.
