### 2021-08-25 #3172 @kthyng
Ok thanks everyone! Looks like we can proceed to the final steps before publication.

@semiautomaticgit Can you verify the version of your software that you want associated with the JOSS publication? Also, please make an archive of the software at a place like Zenodo, making sure to change the metadata so that the title and author list exactly match your JOSS paper.
