### 2021-08-25 #3172 @kthyng
Hi @semiautomaticgit  — in terms of your paper, things are looking good other than the DOIs for your references. Please check those listed above as being invalid and missing, but also look through your references in detail because it looks like they are all generally prefixed with "https://doi.org/" but shouldn't be. Please just include the doi itself. Some have multiple "http" prefixes too. Let me know when you're finished.
