+++ paper/paper.bib
@@ -60,7 +60,7 @@
 	volume = {236},
 	issn = {0034-4257},
 	url = {http://www.sciencedirect.com/science/article/pii/S0034425719304213},
-	doi = {https://doi.org/10.1016/j.rse.2019.111402},
+	doi = {10.1016/j.rse.2019.111402},
 	author = {Weiss, M. and Jacob, F. and Duveiller, G.},
 	year = {2020},
 	pages = {111402},
@@ -71,7 +71,7 @@
 	volume = {122},
 	issn = {0926-5805},
 	url = {http://www.sciencedirect.com/science/article/pii/S0926580520310578},
-	doi = {https://doi.org/10.1016/j.autcon.2020.103477},
+	doi = {10.1016/j.autcon.2020.103477},
 	journal = {Automation in Construction},
 	author = {Garilli, Erika and Bruno, Nazarena and Autelitano, Federico and Roncella, Riccardo and Giuliani, Felice},
 	year = {2021},
@@ -86,7 +86,6 @@
 	journal = {Tropical Conservation Science},
 	author = {Pelage, Latifa and Domalain, Gilles and Lira, Alex S. and Travassos, Paulo and Frédou, Thierry},
 	year = {2019},
-	note = {\_eprint: https://doi.org/10.1177/1940082918822411},
 	pages = {15},
 }
 
@@ -95,7 +94,7 @@
 	volume = {650},
 	issn = {0048-9697},
 	url = {http://www.sciencedirect.com/science/article/pii/S0048969718334053},
-	doi = {https://doi.org/10.1016/j.scitotenv.2018.08.430},
+	doi = {10.1016/j.scitotenv.2018.08.430},
 	journal = {Science of The Total Environment},
 	author = {Huq, Nazmul and Bruns, Antje and Ribbe, Lars},
 	year = {2019},
@@ -224,6 +223,7 @@
 	journal = {Remote Sensing of Environment},
 	author = {Kruse, F. A. and Lefkoff, A. B. and Boardman, J. B. and Heidebrecht, K. B. and Shapiro, A. T. and Barloon, P. J. and Goetz, A. F. H.},
 	year = {1993},
+	doi = {10.1016/0034-4257(93)90013-N},
 	pages = {145--163},
 }
 
@@ -237,6 +237,7 @@
 	author = {Breiman, Leo},
 	month = oct,
 	year = {2001},
+	doi = {10.1023/A:1010933404324 },
 	pages = {5--32},
 }
 
