+++ maininterface/bandcombination.py
@@ -48,19 +48,23 @@ class BandCombination:
 		self.bandSetCombination()
 	
 	# cross classification calculation
-	def bandSetCombination(self, batch = 'No', bandSet = None, rasterOutput = None):
+	def bandSetCombination(self, batch = 'No', bandSetNumber = None, rasterOutput = None):
 		if batch == 'No':
-			combRstPath = cfg.utls.getSaveFileName(None, cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Save band combination raster output'), '', '*.tif', 'tif')
+			combRstPath = cfg.utls.getSaveFileName(None, cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Save band combination raster output'), '', 'TIF file (*.tif);;VRT file (*.vrt)')
 		else:
 			combRstPath = rasterOutput
+		# virtual raster
+		vrtR = 'No'
 		if combRstPath is not False:
-			if combRstPath.lower().endswith('.tif'):
+			if combRstPath.lower().endswith('.vrt'):
+				vrtR = 'Yes'
+			elif combRstPath.lower().endswith('.tif'):
 				pass
 			else:
 				combRstPath = combRstPath + '.tif'
-			if bandSet is None:
+			if bandSetNumber is None:
 				bandSet = cfg.ui.band_set_comb_spinBox.value()
-			bandSetNumber = bandSet - 1
+				bandSetNumber = bandSet - 1
 			if batch == 'No':
 				cfg.uiUtls.addProgressBar()
 			# create list of rasters
@@ -71,7 +75,7 @@ class BandCombination:
 					cfg.uiUtls.removeProgressBar()
 				cfg.mx.msgWar28()
 				# logger
-				cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), " Warning")
+				cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), ' Warning')
 				return 'No'
 			if cfg.bandSetsList[bandSetNumber][0] == 'Yes':
 				ckB = cfg.utls.checkBandSet(bandSetNumber)
@@ -80,37 +84,50 @@ class BandCombination:
 					cfg.uiUtls.removeProgressBar()
 				cfg.mx.msgWar29()
 				# logger
-				cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), " Warning")
+				cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), ' Warning')
 				return 'No'
 			if ckB != 'Yes':
 				pass
-			if len(cfg.bndSetLst) == 0:
+			if len(cfg.bndSetLst) <= 1:
 				if batch == 'No':
 					cfg.uiUtls.removeProgressBar()
 				cfg.mx.msgWar28()
 				# logger
-				cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), " Warning")
+				cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), ' Warning')
 				return 'No'
+			# name of combination field in output table
+			combName = ''
[…]
