+++ README.md
@@ -43,6 +43,7 @@ This plugin is distributed under a GNU General Public License version 3.
 
 ## How to cite 
 Congedo, Luca, (2021). Semi-Automatic Classification Plugin: A Python tool for the download and processing of remote sensing images in QGIS. Journal of Open Source Software, 6(64), 3172, https://doi.org/10.21105/joss.03172
+
 [![DOI](https://joss.theoj.org/papers/10.21105/joss.03172/status.svg)](https://doi.org/10.21105/joss.03172)
 
 ## Code on Zenodo
