+++ ui/ui_semiautomaticclassificationplugin_dock_class.py
@@ -200,6 +200,15 @@ class Ui_DockClass(object):
         self.support_toolButton.setFlat(False)
         self.support_toolButton.setObjectName("support_toolButton")
         self.verticalLayout.addWidget(self.support_toolButton)
+        self.label = QtWidgets.QLabel(self.tab)
+        self.label.setFrameShape(QtWidgets.QFrame.StyledPanel)
+        self.label.setTextFormat(QtCore.Qt.RichText)
+        self.label.setScaledContents(True)
+        self.label.setAlignment(QtCore.Qt.AlignBottom|QtCore.Qt.AlignLeading|QtCore.Qt.AlignLeft)
+        self.label.setWordWrap(True)
+        self.label.setOpenExternalLinks(True)
+        self.label.setObjectName("label")
+        self.verticalLayout.addWidget(self.label)
         self.gridLayout_14.addLayout(self.verticalLayout, 2, 1, 1, 1)
         icon9 = QtGui.QIcon()
         icon9.addPixmap(QtGui.QPixmap(":/plugins/semiautomaticclassificationplugin/icons/semiautomaticclassificationplugin_home.svg"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
@@ -583,6 +592,7 @@ class Ui_DockClass(object):
         self.help_toolButton_2.setText(_translate("DockClass", "Ask a question"))
         self.support_toolButton.setToolTip(_translate("DockClass", "<html><head/><body><p><span style=\" color:#ffffff;\">Support the SCP</span></p></body></html>"))
         self.support_toolButton.setText(_translate("DockClass", "Support the SCP"))
+        self.label.setText(_translate("DockClass", "<html><head/><body><p><span style=\" font-weight:600;\">How to cite</span>:</p><p>Congedo, Luca, (2021). Semi-Automatic Classification Plugin: A Python tool for the download and processing of remote sensing images in QGIS. Journal of Open Source Software, 6(64), 3172, <a href=\"https://doi.org/10.21105/joss.03172\"><span style=\" text-decoration: underline; color:#0000ff;\">https://doi.org/10.21105/joss.03172</span></a></p></body></html>"))
         self.tabWidget_dock.setTabText(self.tabWidget_dock.indexOf(self.tab), _translate("DockClass", "Home"))
         self.button_new_input.setToolTip(_translate("DockClass", "<html><head/><body><p><span >Create a new training input</span></p></body></html>"))
         self.trainingFile_lineEdit.setToolTip(_translate("DockClass", "<html><head/><body><p>Input file path</p></body></html>"))
