+++ i18n/semiautomaticclassificationplugin_de.ts
@@ -5539,7 +5539,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/downloadproducts.py" line="2965"/>
+        <location filename="../maininterface/downloadproducts.py" line="2967"/>
         <source>Searching ...</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6574,7 +6574,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/downloadproducts.py" line="3015"/>
+        <location filename="../maininterface/downloadproducts.py" line="3017"/>
         <source>Downloading</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6589,7 +6589,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/downloadproducts.py" line="3060"/>
+        <location filename="../maininterface/downloadproducts.py" line="3062"/>
         <source>Processing</source>
         <translation type="unfinished"></translation>
     </message>
