+++ i18n/semiautomaticclassificationplugin_fr.ts
@@ -5721,7 +5721,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/crossclassificationTab.py" line="413"/>
+        <location filename="../maininterface/crossclassificationTab.py" line="464"/>
         <source>Classification</source>
         <translation>Classification</translation>
     </message>
@@ -5991,7 +5991,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/crossclassificationTab.py" line="411"/>
+        <location filename="../maininterface/crossclassificationTab.py" line="462"/>
         <source>CROSS MATRIX [</source>
         <translation type="unfinished"></translation>
     </message>
