+++ i18n/semiautomaticclassificationplugin_es.ts
@@ -5373,7 +5373,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation>Guardar matriz de error ráster</translation>
     </message>
     <message>
-        <location filename="../maininterface/crossclassificationTab.py" line="413"/>
+        <location filename="../maininterface/crossclassificationTab.py" line="464"/>
         <source>Classification</source>
         <translation>Classificación</translation>
     </message>
@@ -5988,7 +5988,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/crossclassificationTab.py" line="411"/>
+        <location filename="../maininterface/crossclassificationTab.py" line="462"/>
         <source>CROSS MATRIX [</source>
         <translation type="unfinished"></translation>
     </message>
