+++ i18n/semiautomaticclassificationplugin_pl.ts
@@ -5131,7 +5131,7 @@ ROI</source>
         <translation>Zapisz błąd macierzy wyjścia rastra</translation>
     </message>
     <message>
-        <location filename="../maininterface/crossclassificationTab.py" line="413"/>
+        <location filename="../maininterface/crossclassificationTab.py" line="464"/>
         <source>Classification</source>
         <translation>Klasyfikacja</translation>
     </message>
@@ -5988,7 +5988,7 @@ Chcesz dodać wymagane pola do tego shapefile?</translation>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/crossclassificationTab.py" line="411"/>
+        <location filename="../maininterface/crossclassificationTab.py" line="462"/>
         <source>CROSS MATRIX [</source>
         <translation type="unfinished"></translation>
     </message>
