+++ i18n/semiautomaticclassificationplugin_el_GR.ts
@@ -5489,12 +5489,12 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7321"/>
+        <location filename="../core/utils.py" line="7333"/>
         <source>Build overviews</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../core/utils.py" line="7321"/>
+        <location filename="../core/utils.py" line="7333"/>
         <source>Do you want to build the external overviews of bands?</source>
         <translation type="unfinished"></translation>
     </message>
@@ -5664,7 +5664,7 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="256"/>
+        <location filename="../maininterface/settings.py" line="267"/>
         <source>Transparency </source>
         <translation type="unfinished"></translation>
     </message>
@@ -5674,52 +5674,52 @@ Do you want to add the required fields to this shapefile?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="196"/>
+        <location filename="../maininterface/settings.py" line="207"/>
         <source>Reset field names</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="196"/>
+        <location filename="../maininterface/settings.py" line="207"/>
         <source>Are you sure you want to reset field names?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="207"/>
+        <location filename="../maininterface/settings.py" line="218"/>
         <source>Reset variable name</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="207"/>
+        <location filename="../maininterface/settings.py" line="218"/>
         <source>Are you sure you want to reset variable name?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="215"/>
+        <location filename="../maininterface/settings.py" line="226"/>
         <source>Reset group name</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="215"/>
+        <location filename="../maininterface/settings.py" line="226"/>
         <source>Are you sure you want to reset group name?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="223"/>
+        <location filename="../maininterface/settings.py" line="234"/>
         <source>Change temporary directory</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="223"/>
+        <location filename="../maininterface/settings.py" line="234"/>
         <source>Are you sure you want to change the temporary directory?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="241"/>
+        <location filename="../maininterface/settings.py" line="252"/>
         <source>Reset temporary directory</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../maininterface/settings.py" line="241"/>
+        <location filename="../maininterface/settings.py" line="252"/>
         <source>Are you sure you want to reset the temporary directory?</source>
         <translation type="unfinished"></translation>
     </message>
[…]
