+++ maininterface/bandsetTab.py
@@ -249,23 +249,45 @@ class BandsetTab:
 			bandSetNumber = cfg.ui.Band_set_tabWidget.currentIndex()
 		tW = eval('cfg.ui.tableWidget__' + cfg.bndSetTabList[bandSetNumber])
 		satellite = None
-		if date is None:
-			date = ''
 		if batch == 'No':
 			files = cfg.utls.getOpenFileNames(None , cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Select a raster'), '', 'Raster (*.*)')
 		else:
 			cfg.utls.clearTable(tW)
 			fileList = fileListString.split(',')
 			files = []
-			if len(fileList) == 2 and cfg.QDirSCP(fileList[0].strip("'")).exists():
+			# directory
+			if len(fileList) <= 2 and cfg.QDirSCP(fileList[0].strip("'")).exists():
 				dirF = fileList[0].strip("'").rstrip('/')
-				filter = fileList[1].strip("'").strip()
+				try:
+					filter = fileList[1].strip("'").strip()
+				except:
+					filter = ''
 				for r, d, f in cfg.osSCP.walk(dirF):
 					for x in f:
 						if filter in x:
 							files.append(cfg.osSCP.path.join(r, x))
 				files = sorted(files)
-			else:				
+				if date is not None:
+					if date.lower() == 'auto':
+						# date from directory name
+						try:
+							dStr = cfg.datetimeSCP.datetime.strptime(dirF[-10:], '%Y-%m-%d')
+							date = dirF[-10:]
+						except:
+							try:
+								dirPart = dirF.split('_')
+								for dP in dirPart:
+									dPP = dP.lower().split('t')[0]
+									try:
+										dStr = cfg.datetimeSCP.datetime.strptime(dPP, '%Y%m%d')
+										dPPS = dStr.strftime('%Y-%m-%d')
+										date = dPPS
+										break
+									except:
+										pass
+							except:
+								pass
+			else:
 				for f in fileList:
 					files.append(f.strip())
 			try:
@@ -286,6 +308,8 @@ class BandsetTab:
 				additiveFactor = additiveFactorString.split(',')
 			except:
 				additiveFactor = None
+			if date is None:
+				date = ''
 		cfg.BandTabEdited = 'No'
 		tW.blockSignals(True)
 		# count table rows
@@ -301,11 +325,7 @@ class BandsetTab:
 			if iBC > 1:
 				r =cfg.utls.addRasterLayer(files[0])
 				cfg.utls.clearTable(tW)
-				cfg.QtWidgetsSCP.qApp.processEvents()
-				cfg.ipt.checkRefreshRasterLayer()
-				# load project image name in combo
-				id = cfg.ui.image_raster_name_combo.findText(r.name())
-				cfg.ui.image_raster_name_combo.setCurrentIndex(id)
+				self.rasterToBandName(r.name(), date = date, bandSetNumber = bandSetNumber)
 				return 'Yes'
 		# check if single raster
 		if c > 0 and cfg.bandSetsList[bandSetNumber][0] == 'No':
@@ -316,7 +336,7 @@ class BandsetTab:
 				iBC = cfg.utls.getNumberBandRaster(i)
 			except Exception as err:
 				# logger
[…]
