+++ maininterface/stackrasterbands.py
@@ -64,7 +64,7 @@ class StackRasterBands:
 					cfg.utls.logCondition(str(__name__) + '-' + str(cfg.inspectSCP.stack()[0][3])+ ' ' + cfg.utls.lineOfCode(), ' Warning')
 					return 'No'
 				if outputFile is None:
-					rstrOut = cfg.utls.getSaveFileName(None , cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Save raster'), '', '*.tif', 'tif')
+					rstrOut = cfg.utls.getSaveFileName(None , cfg.QtWidgetsSCP.QApplication.translate('semiautomaticclassificationplugin', 'Save raster'), '', 'TIF file (*.tif);;VRT file (*.vrt)')
 					if rstrOut is False:
 						if batch == 'No':
 							cfg.uiUtls.removeProgressBar()
@@ -77,15 +77,18 @@ class StackRasterBands:
 				cfg.uiUtls.removeProgressBar()
 			return 'No'
 		if rstrOut is not False:
-			if rstrOut.lower().endswith('.tif'):
-				pass
+			if rstrOut.lower().endswith('.vrt'):
+				outputFormat = 'vrt'
+			elif rstrOut.lower().endswith('.tif'):
+				outputFormat = 'GTiff'
 			else:
 				rstrOut = rstrOut + '.tif'
+				outputFormat = 'GTiff'
 			if outputFile is None:
 				cfg.uiUtls.addProgressBar()
 			cfg.uiUtls.updateBar(10)
-			cfg.utls.makeDirectory(cfg.osSCP.path.dirname(outputFile))
-			st = cfg.utls.mergeRasterBands(cfg.bndSetLst, rstrOut, compress = 'Yes')
+			cfg.utls.makeDirectory(cfg.osSCP.path.dirname(rstrOut))
+			st = cfg.utls.mergeRasterBands(cfg.bndSetLst, rstrOut, compress = 'Yes', outFormat = outputFormat)
 			if cfg.osSCP.path.isfile(rstrOut):
 				cfg.cnvs.setRenderFlag(False)
 				cfg.utls.addRasterLayer(rstrOut)
